Documentation initial sweep
Added exp_lists, exp_maps, exp_hll, exp_bit to properly list documentation, various other documentation changes.
