+++ lib/aerospike.js
@@ -138,12 +138,12 @@ exports.bitwise = require('./bitwise')
 exports.operations = require('./operations')
 
 /**
- * The {@link module:aerospike/exp_operations|ExpOperation} module provides functions
+ * The {@link module:aerospike/exp/operations|ExpOperation} module provides functions
  * to create operations using <a href="https://docs.aerospike.com/guide/expressions">&uArr;Aerospike Expressions</a>
  * for the {@link Client#operate} command.
  *
  *
- * @summary {@link module:aerospike/exp_operations|aerospike/exp_operations} module
+ * @summary {@link module:aerospike/exp/operations|aerospike/exp/operations} module
  */
 exp.operations = require('./exp_operations')
 
@@ -205,9 +205,8 @@ exports.features = require('./features')
 exports.AerospikeError = AerospikeError
 
 /**
- * The main interface of the Aerospike client. Through the Client class,
- * commands such as put, get or query can be sent to an Aerospike database
- * cluster.
+ * The main interface of the Aerospike client. Commands such as put, get or query
+ * can be sent to an Aerospike database cluster using the Client class.
  *
  * @summary {@link Client} class
  */
@@ -352,18 +351,25 @@ exports.language = as.language
  * @example <caption>Setting log level for new client</caption>
  *
  * const Aerospike = require('aerospike')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
  * var config = {
- *   log: { level: Aerospike.log.INFO }
+ *   hosts: '192.168.33.10:3000',
+ *   log: { level: Aerospike.log.DEBUG },
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
  * }
  * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
- *
  *   var key = new Aerospike.Key('test', 'demo', 'k1')
- *   client.get(key, (error, record) => {
+ *   client.put(key, {example: 31}, (error) => {
  *     if (error) throw error
- *     console.info(record)
- *     client.close()
+ *     client.get(key, (error, record) => {
+ *       console.info(record)
+ *       client.close()
+ *     })
  *   })
  * })
  */
@@ -388,14 +394,22 @@ exports.log = as.log
  *
  * const Aerospike = require('aerospike')
  * const Key = Aerospike.Key
- *
- * Aerospike.connect((error, client) => {
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *   let key = new Key('test', 'demo', 'key1')
  *   let bins = { 'int': 1 }
  *   let meta = { ttl: 86400 }
  *
- *   client.put(key, record, meta, (error) => {
+ *   client.put(key, bins, meta, (error) => {
  *     if (error) throw error
  *
  *     bins = { 'int': 2 }
@@ -452,13 +466,13 @@ exports.indexDataType = as.indexDataType
  * containing scalar values (i.e. integer, string).
  * @property {number} LIST - SI for bins containing
  * <a href="http://www.aerospike.com/docs/guide/cdt-list.html" title="Aerospike List Data Type">&uArr;Lists</a>;
- * the index will be build over the individual entries of the list.
+ * The index will be built over the individual entries of the list.
  * @property {number} MAPKEYS - SI for bins containing
  * <a href="http://www.aerospike.com/docs/guide/cdt-map.html" title="Aerospike Maps Data Type">&uArr;Maps</a>;
- * the index will be build over the individual keys of the map entries.
+ * The index will be built over the individual keys of the map entries.
  * @property {number} MAPVALUES - SI for bins containing
  * <a href="http://www.aerospike.com/docs/guide/cdt-map.html" title="Aerospike Maps Data Type">&uArr;Maps</a>;
- * the index will be build over the individual values of the map entries.
+ * The index will be built over the individual values of the map entries.
  *
  * @see {@link Client#createIndex}
  */
@@ -485,47 +499,71 @@ exports.client = function (config) {
 /**
  * Creates a new {@link Client} instance and connects to the Aerospike cluster.
  *
- * Calling `Aerospike.connect(config, (err, client) -> { ... })` is equivalent to calling
- *
- *     const client = Aerospike.client(config)
- *     client.connect((err) -> { ... })
- *
  * @param {Config} [config] - The configuration for the client.
  * @param {connectCallback} [callback] - The function to call, once the client is connected to the cluster successfully.
  *
  * @return {?Promise} If no callback function is passed, the function returns
  * a Promise resolving to the connected client.
+ * @example <caption>Connection can be established using the {@link module:aerospike|aerospike} module.</caption>
+ * 
+ * const Aerospike = require('aerospike')
  *
- * @example <caption>Using callback function</caption>
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * 
+ * Aerospike.connect(config, (err, client) => { 
+ *   console.log("Connected. Closing now.")
+ *   client.close()
+ * })
+ * 
+ * @example <caption>Connection can also be established using the {@link Client} module.</caption>
+ * 
+ * const Aerospike = require('aerospike')
+ *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ *  
+ * const client = Aerospike.client(config)
+ * client.connect((err) => { 
+ *   console.log("Connected. Closing now.")
+ *   client.close()
+ * })
+ * 
+ * @example <caption>A connection established using callback function.</caption>
  *
  * const Aerospike = require('aerospike')
  *
- * let config = { ... }
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
  * Aerospike.connect(config, (error, client) => {
  *   if (error) {
  *     console.error('Failed to connect to cluster: %s', error.message)
  *     process.exit()
  *   } else {
  *     // client is ready to accept commands
+ *     console.log("Connected. Now closing connection.")
  *     client.close()
  *   }
  * })
  *
- * @example <caption>Returning a Promise</caption>
+ * @example <caption>A connection established by returning a Promise.</caption>
  *
  * const Aerospike = require('aerospike')
- *
- * let config = { ... }
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
  * Aerospike.connect(config)
  *   .then(client => {
  *     // client is ready to accept commands
- *     return client.get(key)
- *       .then(record => console.info(record))
- *       .then(() => client.close())
- *       .catch(error => {
- *         console.error('Failed to get record: %s', error.message)
- *         client.close()
- *       })
+ *     console.log("Connected. Now Closing Connection.")
+ *     client.close()
  *   })
  *   .catch(error => {
  *     console.error('Failed to connect to cluster: %s', error.message)
@@ -568,8 +606,24 @@ exports.connect = function (config, callback) {
  *
  * const Aerospike = require('aerospike')
  *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * 
  * Aerospike.setDefaultLogging({
  *   level: Aerospike.log.TRACE
+ * })
+ * 
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) {
+ *     console.error('Failed to connect to cluster: %s', error.message)
+ *     process.exit()
+ *   } else {
+ *     // client is ready to accept commands
+ *     console.log("Connected. Now closing connection.")
+ *     client.close()
+ *   }
  * })
  */
 exports.setDefaultLogging = function (logInfo) {
@@ -582,7 +636,7 @@ exports.setDefaultLogging = function (logInfo) {
  * @description Note that there is only one instance of the command queue that
  * is shared by all client instances, even client instances connected to
  * different Aerospike clusters. The <code>setupGlobalCommandQueue</code>
- * method must be called before any clien} instances are connected.
+ * method must be called before any client instances are connected.
  *
  * @param {CommandQueuePolicy} policy - Set of policy values governing the
  * behaviour of the global command queue.
@@ -594,11 +648,13 @@ exports.setupGlobalCommandQueue = function (policy) {
   EventLoop.setCommandQueuePolicy(policy)
 }
 
+
 /**
-  * @summary This module {@link module:aerospike/batchType|aerospike/batch_type}
-  * contains a list of batch record types.
-  *
-  */
+ * The {@link module:aerospike/batchType|aerospike/batch_type}
+ * module contains a list of batch record types.
+ *
+ * @summary {@link module:aerospike/batchType|aerospike/batch_type} module
+ */
 exports.batchType = require('./batch_type')
 
 // Set default log level
