+++ lib/batch_type.js
@@ -22,7 +22,7 @@ const batchType = as.batchTypes
 /**
  * @module aerospike/batchType
  *
- * @description batch record type.
+ * @description Identifies batch record type with designated enumerated type 
  */
 
 // ========================================================================
@@ -30,26 +30,28 @@ const batchType = as.batchTypes
 // ========================================================================
 module.exports = {
   /**
-     * An instance of class Record that is used in a batch for read operations.
+     * An instance of class {@link Record} that is used in a batch for read operations.
      * @const {number}
      */
   BATCH_READ: batchType.BATCH_READ,
 
   /**
-     * An instance of class Record that is used in a batch for write operations.
+     * An instance of class {@link Record} that is used in a batch for write operations.
      * @const {number}
      */
   BATCH_WRITE: batchType.BATCH_WRITE,
 
   /**
-     * An instance of class Record that is used in a batch for applying record.
+     * An instance of class {@link Record} that is used in a batch for applying record.
      * @const {number}
      */
   BATCH_APPLY: batchType.BATCH_APPLY,
 
   /**
-     * An instance of class Record that is used in a batch for removal operations.
+     * An instance of class {@link Record} that is used in a batch for removal operations.
      * @const {number}
      */
   BATCH_REMOVE: batchType.BATCH_REMOVE
+
+
 }
