+++ lib/bitwise.js
@@ -42,7 +42,7 @@ const as = require('bindings')('aerospike.node')
 const opcodes = as.bitOperations
 
 /**
- * @class module:aerospike/bitwise~BitwiseOperation
+ * @class aerospike/bitwise~BitwiseOperation
  *
  * @classdesc Use the methods in the {@link module:aerospike/bitwise|bitwise}
  * module to create bitwise operations for use with the {@link Client#operate}
@@ -61,7 +61,7 @@ class BitwiseOperation extends Operation {
 }
 
 /**
- * @class module:aerospike/bitwise~OverflowableBitwiseOp
+ * @class aerospike/bitwise~OverflowableBitwiseOp
  *
  * @classdesc Bitwise operation variant that can overflow/underflow.
  *
