+++ lib/cdt_context.js
@@ -93,7 +93,7 @@ class CdtContext {
   }
 
   /**
-   * @summary yLookup map by index offset.
+   * @summary Lookup map by index offset.
    *
    * @description If the index is negative, the resolved index starts backwards
    * from end of list. If an index is out of bounds, a parameter error will be
