+++ lib/client.js
@@ -60,16 +60,21 @@ function eventsCallback (event) {
    * @example
    *
    * const Aerospike = require('aerospike')
-   *
-   * Aerospike.connect((error, client) => {
+   * 
+   * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+   * var config = {
+   *   hosts: '192.168.33.10:3000',
+   * }
+   * 
+   * Aerospike.connect(config, (error, client) => {
    *   if (error) throw error
    *
    *   client.on('disconnected', () => {
-   *     console.warn('Client got disconnected from cluster')
+   *     console.log('Client got disconnected from cluster')
    *   })
    *
-   *  // client is now ready to accept commands, e.g. get/put/...
-   *
+   *   // client is now ready to accept commands, e.g. get/put/...
+   *   client.close()
    * })
    */
   this.emit(event.name, event)
@@ -98,7 +103,7 @@ function eventsCallback (event) {
    *
    *   client.on('event', (event) => {
    *     var now = new Date().toUTCString()
-   *     console.info(now, event.name, event.nodeName)  // Example output:
+   *     console.log(now, event.name, event.nodeName)  // Example output:
    *                  // Thu, 13 Jul 2017 06:47:35 GMT nodeAdded BB94DC07D270009
    *                  // Thu, 13 Jul 2017 06:47:35 GMT nodeAdded C1D4DC0AD270002
    *                  // Thu, 13 Jul 2017 06:48:52 GMT nodeRemoved C1D4DC0AD270002
@@ -106,8 +111,8 @@ function eventsCallback (event) {
    *                  // Thu, 13 Jul 2017 06:49:08 GMT disconnected
    *   })
    *
-   *  // client is now ready to accept commands, e.g. get/put/...
-   *
+   *   // client is now ready to accept commands, e.g. get/put/...
+   *   client.close()
    * })
    */
   this.emit('event', event)
@@ -188,10 +193,15 @@ Client.prototype.asExec = function (cmd, args) {
  * @example
  *
  * const Aerospike = require('aerospike')
- * Aerospike.connect((error, client) => {
- *   if (err) throw err
- *   console.log(client.getNodes()) // [ { name: 'BB9EB63B2005452', address: '127.0.0.1:3000' },
- *                                  //   { name: 'C1DEB63B2005452', address: '127.0.0.1:3100' } ]
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   console.log(client.getNodes()) // [ { name: 'SAMPLEADDRESS', address: 'SAMPLENAME' }, ...]
  *   client.close()
  * })
  *
@@ -247,31 +257,46 @@ Client.prototype.removeSeedHost = function (hostname, port) {
  *
  * @example
  *
+ * const Aerospike = require('aerospike')
+ * const Key = Aerospike.Key
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     batch : new Aerospike.BatchPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *   }
+ * }
+ * 
  * var keys = [
  *   new Key('test', 'demo', 'key1'),
  *   new Key('test', 'demo', 'key2'),
  *   new Key('test', 'demo', 'key3')
  * ]
- *
- * client.batchExists(keys, function (error, results) {
- *   if (error) {
- *     // handle failure
- *   } else {
- *     results.forEach(function (result) {
+ * 
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   client.batchExists(keys, (error, results) => {
+ *     if (error) throw error
+ *     results.forEach((result) => {
  *       switch (result.status) {
- *         case status.OK:
- *           // record found
+ *         case Aerospike.status.OK:
+ *           console.log("Record found")
  *           break
- *         case status.ERR_RECORD_NOT_FOUND:
- *           // record not found
+ *         case Aerospike.status.ERR_RECORD_NOT_FOUND:
+ *           console.log("Record not found")
  *           break
  *         default:
  *           // error while reading record
+ *           console.log("Other error")
  *           break
  *       }
  *     })
- *   }
+ *     client.close()
+ *   })
  * })
+ * 
  */
 Client.prototype.batchExists = function (keys, policy, callback) {
   if (typeof policy === 'function') {
@@ -300,31 +325,46 @@ Client.prototype.batchExists = function (keys, policy, callback) {
  *
  * @example
  *
+ * const Aerospike = require('aerospike')
+ * const Key = Aerospike.Key
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     batch : new Aerospike.BatchPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *   }
+ * }
+ * 
  * var keys = [
  *   new Key('test', 'demo', 'key1'),
  *   new Key('test', 'demo', 'key2'),
  *   new Key('test', 'demo', 'key3')
  * ]
- *
- * client.batchGet(keys, function (error, results) {
- *   if (error) {
- *     // handle failure
- *   } else {
- *     results.forEach(function (result) {
+ * 
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   client.batchGet(keys, (error, results) => {
+ *     if (error) throw error
+ *     results.forEach((result) => {
  *       switch (result.status) {
- *         case status.OK:
- *           // record found - bin values are available in result.record
+ *         case Aerospike.status.OK:
+ *           console.log("Record found")
  *           break
- *         case status.ERR_RECORD_NOT_FOUND:
- *           // record not found
+ *         case Aerospike.status.ERR_RECORD_NOT_FOUND:
+ *           console.log("Record not found")
  *           break
  *         default:
  *           // error while reading record
+ *           console.log("Other error")
  *           break
  *       }
  *     })
- *   }
+ *     client.close()
+ *   })
  * })
+ * 
  */
 Client.prototype.batchGet = function (keys, policy, callback) {
   if (typeof policy === 'function') {
@@ -365,8 +405,19 @@ Client.prototype.batchGet = function (keys, policy, callback) {
  * @since v2.0
  *
  * @example
- *
  * const Aerospike = require('aerospike')
+ * const batchType = Aerospike.batchType
+ * const op = Aerospike.operations
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     batch : new Aerospike.BatchPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *   }
+ * }
+ * 
  * var batchRecords = [
  *   { type: batchType.BATCH_READ,
  *     key: new Aerospike.Key('test', 'demo', 'key1'), bins: ['i', 's'] },
@@ -378,13 +429,25 @@ Client.prototype.batchGet = function (keys, policy, callback) {
  *          op.read('blob-bin')
  *         ]}
  * ]
- * Aerospike.connect(function (error, client) {
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
- *   client.batchRead(batchRecords, function (error, results) {
+ *   client.batchRead(batchRecords, (error, results) => {
  *     if (error) throw error
- *     results.forEach(function (result) {
- *       console.log(result)
+ *     results.forEach((result) => {
+ *       switch (result.status) {
+ *         case Aerospike.status.OK:
+ *           console.log("Record found")
+ *           break
+ *         case Aerospike.status.ERR_RECORD_NOT_FOUND:
+ *           console.log("Record not found")
+ *           break
+ *         default:
+ *           // error while reading record
+ *           console.log("Other error")
+ *           break
+ *       }
  *     })
+ *     client.close()
  *   })
  * })
  */
@@ -423,6 +486,19 @@ Client.prototype.batchRead = function (records, policy, callback) {
  * @example
  * const Aerospike = require('aerospike')
  * const batchType = Aerospike.batchType
+ * const op = Aerospike.operations
+ * const GeoJSON = Aerospike.GeoJSON
+ * const exp = Aerospike.exp
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     batch : new Aerospike.BatchPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *   }
+ * }
+ * 
  * var batchRecords = [
  *   { type: batchType.BATCH_READ,
  *     key: new Aerospike.Key('test', 'demo', 'key1'),
@@ -448,7 +524,7 @@ Client.prototype.batchRead = function (records, policy, callback) {
  *         gen: Aerospike.policy.gen.EQ,
  *         exists: Aerospike.policy.exists.CREATE,
  *         durableDelete: true
- *       })},
+ *       }),
  *   { type: batchType.BATCH_REMOVE,
  *     key: new Aerospike.Key('test', 'demo', 'key5'),
  *     policy: new Aerospike.BatchRemovePolicy({
@@ -457,16 +533,29 @@ Client.prototype.batchRead = function (records, policy, callback) {
  *         commitLevel: Aerospike.policy.commitLevel.ALL,
  *         gen: Aerospike.policy.gen.EQ,
  *         durableDelete: true
- *       })},
+ *       }),
  *    }
  * ]
- * Aerospike.connect(function (error, client) {
+ * 
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
- *   client.batchWrite(batchRecords, function (error, results) {
+ *   client.batchWrite(batchRecords, (error, results) => {
  *     if (error) throw error
- *     results.forEach(function (result) {
- *       console.log(result)
+ *     results.forEach((result) => {
+ *       switch (result.status) {
+ *         case Aerospike.status.OK:
+ *           console.log("Record found")
+ *           break
+ *         case Aerospike.status.ERR_RECORD_NOT_FOUND:
+ *           console.log("Record not found")
+ *           break
+ *         default:
+ *           // error while reading record
+ *           console.log("Other error")
+ *           break
+ *       }
  *     })
+ *     client.close()
  *   })
  * })
  */
@@ -506,6 +595,15 @@ Client.prototype.batchWrite = function (records, policy, callback) {
  *
  * @example
  * const Aerospike = require('aerospike')
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROSPIKE SERVER NODE HERE!
+ * const config = {  
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     batchApply : new Aerospike.BatchApplyPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *   }
+ * }
+ * 
  * const batchType = Aerospike.batchType
  * var batchRecords = [
  *   { type: batchType.BATCH_READ,
@@ -543,9 +641,9 @@ Client.prototype.batchWrite = function (records, policy, callback) {
  *          }
  *    }
  * ]
- * Aerospike.connect(function (error, client) {
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
- *   client.batchApply(batchRecords, udf, function (error, results) {
+ *   client.batchApply(batchRecords, (error, results) => {
  *     if (error) throw error
  *     results.forEach(function (result) {
  *       console.log(result)
@@ -595,6 +693,16 @@ Client.prototype.batchApply = function (records, udf, batchPolicy, batchApplyPol
  * @example
  * const Aerospike = require('aerospike')
  * const batchType = Aerospike.batchType
+ * const exp = Aerospike.exp
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     batch : new Aerospike.BatchPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *   }
+ * }
  * var batchRecords = [
  *   { type: batchType.BATCH_REMOVE,
  *     key: new Aerospike.Key('test', 'demo', 'key5'),
@@ -604,7 +712,7 @@ Client.prototype.batchApply = function (records, udf, batchPolicy, batchApplyPol
  *         commitLevel: Aerospike.policy.commitLevel.ALL,
  *         gen: Aerospike.policy.gen.EQ,
  *         durableDelete: true
- *       })},
+ *       }),
  *    },
  *   { type: batchType.BATCH_REMOVE,
  *     key: new Aerospike.Key('test', 'demo', 'key6'),
@@ -614,16 +722,29 @@ Client.prototype.batchApply = function (records, udf, batchPolicy, batchApplyPol
  *         commitLevel: Aerospike.policy.commitLevel.ALL,
  *         gen: Aerospike.policy.gen.EQ,
  *         durableDelete: true
- *       })},
+ *       }),
  *    }
  * ]
- * Aerospike.connect(function (error, client) {
+ * 
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
- *   client.batchRemove(batchRecords, function (error, results) {
+ *   client.batchRemove(batchRecords, (error, results) => {
  *     if (error) throw error
- *     results.forEach(function (result) {
- *       console.log(result)
+ *     results.forEach((result) => {
+ *       switch (result.status) {
+ *         case Aerospike.status.OK:
+ *           console.log("Record found")
+ *           break
+ *         case Aerospike.status.ERR_RECORD_NOT_FOUND:
+ *           console.log("Record not found")
+ *           break
+ *         default:
+ *           // error while reading record
+ *           console.log("Other error")
+ *           break
+ *       }
  *     })
+ *     client.close()
  *   })
  * })
  */
@@ -660,32 +781,46 @@ Client.prototype.batchRemove = function (records, batchPolicy, batchRemovePolicy
  * @deprecated since v2.0 - use {@link Client#batchRead} instead.
  *
  * @example
- *
+ * const Aerospike = require('aerospike')
+ * const Key = Aerospike.Key
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     batch : new Aerospike.BatchPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *   }
+ * }
+ * 
  * var keys = [
  *   new Key('test', 'demo', 'key1'),
  *   new Key('test', 'demo', 'key2'),
  *   new Key('test', 'demo', 'key3')
  * ]
+ * 
  * var bins = ['s', 'i']
- *
- * client.batchSelect(keys, bins, function (error, results) {
- *   if (error) {
- *     // handle failure
- *   } else {
- *     results.forEach(function (result) {
+ * 
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   client.batchSelect(keys, bins, (error, results) => {
+ *     if (error) throw error
+ *     results.forEach((result) => {
  *       switch (result.status) {
- *         case status.OK:
- *           // record found - selected bins are available in result.record.
+ *         case Aerospike.status.OK:
+ *           console.log("Record found")
  *           break
- *         case status.ERR_RECORD_NOT_FOUND:
- *           // record not found
+ *         case Aerospike.status.ERR_RECORD_NOT_FOUND:
+ *           console.log("Record not found")
  *           break
  *         default:
  *           // error while reading record
+ *           console.log("Other error")
  *           break
  *       }
  *     })
- *   }
+ *     client.close()
+ *   })
  * })
  */
 Client.prototype.batchSelect = function (keys, bins, policy, callback) {
@@ -710,11 +845,21 @@ Client.prototype.batchSelect = function (keys, bins, policy, callback) {
  * @example
  *
  * const Aerospike = require('aerospike')
- * Aerospike.connect(function (error, client) {
- *   if (error) throw error
- *   // client ready to receive commands
- *   client.close()
- * })
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * Aerospike.connect(config)
+ *   .then(client => {
+ *     // client is ready to accept commands
+ *     console.log("Connected. Now Closing Connection.")
+ *     client.close()
+ *   })
+ *   .catch(error => {
+  *    client.close()
+ *     console.error('Failed to connect to cluster: %s', error.message)
+ *   })
  */
 Client.prototype.close = function (releaseEventLoop = false) {
   if (this.isConnected(false)) {
@@ -759,18 +904,22 @@ Client.prototype.close = function (releaseEventLoop = false) {
  * @see {@link module:aerospike.connect} - Initial host connection timeout in milliseconds.
  * @see module:aerospike.releaseEventLoop
  *
- * @example <caption>Using callback function</caption>
+ * @example <caption>A connection established using callback function.</caption>
  *
  * const Aerospike = require('aerospike')
  *
- * var config = { ... }
- * var client = Aerospike.client(config)
- * client.connect((error) => {
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) {
  *     console.error('Failed to connect to cluster: %s', error.message)
  *     process.exit()
  *   } else {
  *     // client is ready to accept commands
+ *     console.log("Connected. Now closing connection.")
+ *     client.close()
  *   }
  * })
  */
@@ -861,7 +1010,13 @@ Client.prototype.connect = function (callback) {
  * @example
  *
  * const Aerospike = require('aerospike')
- * Aerospike.connect((error, client) => {
+ *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *
  *   // create index over user's recent locations
@@ -941,7 +1096,13 @@ Client.prototype.createIndex = function (options, policy, callback) {
  * @example
  *
  * const Aerospike = require('aerospike')
- * Aerospike.connect((error, client) => {
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *
  *   var binName = 'age'
@@ -991,7 +1152,13 @@ Client.prototype.createIntegerIndex = function (options, policy, callback) {
  * @example
  *
  * const Aerospike = require('aerospike')
- * Aerospike.connect((error, client) => {
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *
  *   var binName = 'name'
@@ -1041,7 +1208,12 @@ Client.prototype.createStringIndex = function (options, policy, callback) {
  * @example
  *
  * const Aerospike = require('aerospike')
- * Aerospike.connect((error, client) => {
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *
  *   var binName = 'location'
@@ -1094,20 +1266,33 @@ Client.prototype.createGeo2DSphereIndex = function (options, policy, callback) {
  * @see {@link Query#background} and {@link Scan#background} to apply a Record UDF function to multiple records instead.
  *
  * @example
- *
- * var key = new Key('test', 'demo', value')
+ * 
+ * const Aerospike = require('aerospike')
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROSPIKE SERVER NODE HERE!
+ * const config = {  
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     apply : new Aerospike.ApplyPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *   }
+ * }
+ * 
+ * var key = new Aerospike.Key('test', 'demo', 'value')
+ * 
  * var udfArgs = {
  *   module: 'my_udf_module',
  *   funcname: 'my_udf_function',
  *   args: ['abc', 123, 4.5]
  * }
  *
- * client.apply(key, udfArgs, (error, result) => {
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
- *
- *   console.log('Result of calling my_udf_function:', result)
+ *   client.apply(key, udfArgs, (error, result) => {
+ *     if (error) throw error
+ *  
+ *     console.log('Result of calling my_udf_function:', result)
+ *   })
  * })
- *
  */
 Client.prototype.apply = function (key, udfArgs, policy, callback) {
   if (typeof policy === 'function') {
@@ -1137,9 +1322,16 @@ Client.prototype.apply = function (key, udfArgs, policy, callback) {
  * @example
  *
  * const Aerospike = require('aerospike')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *   }
+ * }
+ * 
  * let key = new Aerospike.Key('test', 'demo', 'key1')
- * Aerospike.connect()
+ * Aerospike.connect(config)
  *   .then(client => {
  *     return client.exists(key)
  *       .then(exists => console.info('Key "%s" exists: %s', key.key, exists))
@@ -1178,16 +1370,26 @@ Client.prototype.exists = function exists (key, policy, callback) {
  * a <code>Promise</code> that resolves to a {@link Record}.
  *
  * @example
+ * const Aerospike = require('aerospike')
+ * var key = new Aerospike.Key('test', 'demo', 'key1')
  *
- * var key = new Key('test', 'demo', 'key1')
- *
- * client.get(key, (error, record) => {
- *   if (error) {
- *     // handle failure
- *   } else {
- *     // handle success
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
  *   }
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   client.get(key, (error, record) => {
+ *     if (error) throw error
+ *     console.log(record)
+ *     client.close()
+ *   })
  * })
+ * 
  */
 Client.prototype.get = function (key, policy, callback) {
   if (typeof policy === 'function') {
@@ -1215,13 +1417,18 @@ Client.prototype.get = function (key, policy, callback) {
  *
  * @example
  *
- * client.indexRemove('test', 'index', function (error) {
- *   if (error) {
- *     // handle failure
- *   } else {
- *     // handle success
- *   }
- * });
+ * const Aerospike = require('aerospike')
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * Aerospike.connect(config, (error, client) => {
+ *   client.indexRemove('location', 'locationIndex', (error) => {
+ *     if (error) throw error
+ *     client.close()
+ *   })
+ * })
  */
 Client.prototype.indexRemove = function (namespace, index, policy, callback) {
   if (typeof policy === 'function') {
@@ -1258,14 +1465,23 @@ Client.prototype.indexRemove = function (namespace, index, policy, callback) {
  * @deprecated since v3.11.0 - use {@link Client#infoNode} or {@link Client#infoAny} instead.
  *
  * @example <caption>Sending a 'statistics' info query to a single host</caption>
- *
- * client.info('statistics', {addr: '127.0.0.1', port: 3000}, function (error, response) {
- *   if (error) {
- *     // handle failure
- *   } else {
- *     // handle success
- *   }
+ * 
+ * const Aerospike = require('aerospike')
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   client.info('statistics', {addr: '192.168.33.10', port: 3000}, (error, response) => {
+ *     if (error) throw error
+ *     console.log(response)
+ *     client.close() 
+ *   })
  * })
+ * 
  */
 Client.prototype.info = function (request, host, policy, callback) {
   if (typeof policy === 'function') {
@@ -1302,13 +1518,22 @@ Client.prototype.info = function (request, host, policy, callback) {
  *
  * @example <caption>Sending 'statistics' info command to random cluster node</caption>
  *
- * client.infoAny('statistics', function (error, response) {
- *   if (error) {
- *     // handle failure
- *   } else {
- *     // handle success
- *   }
+ * const Aerospike = require('aerospike')
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   client.infoAny('statistics', (error, response) => {
+ *     if (error) throw error
+ *     console.log(response)
+ *     client.close() 
+ *   })
  * })
+ * 
  */
 Client.prototype.infoAny = function (request, policy, callback) {
   if (typeof request === 'function') {
@@ -1346,15 +1571,22 @@ Client.prototype.infoAny = function (request, policy, callback) {
  *
  * @example <caption>Sending info command to whole cluster</caption>
  *
- * client.infoAll('statistics', function (error, responses) {
- *   if (error) {
- *     // handle failure
- *   } else {
- *     responses.forEach((function (info) {
- *       // handle response
- *     })
- *   }
+ * const Aerospike = require('aerospike')
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   client.infoAll('statistics', (error, response) => {
+ *     if (error) throw error
+ *     console.log(response)
+ *     client.close() 
+ *   })
  * })
+ * 
  */
 Client.prototype.infoAll = function (request, policy, callback) {
   if (typeof request === 'function') {
@@ -1393,10 +1625,24 @@ Client.prototype.infoAll = function (request, policy, callback) {
  *
  * @example <caption>Sending 'statistics' info command to specific cluster node</caption>
  *
- * const node = client.getNodes().pop()
- * client.infoNode('statistics', node).then(info => {
- *   // process info
+ * 
+ * const Aerospike = require('aerospike')
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   const node = client.getNodes().pop()
+ *   client.infoNode('statistics', node, (error, response) => {
+ *     if (error) throw error
+ *     console.log(response)
+ *     client.close() 
+ *   })
  * })
+ * 
  */
 Client.prototype.infoNode = function (request, node, policy, callback) {
   if (typeof policy === 'function') {
@@ -1456,24 +1702,34 @@ Client.prototype.isConnected = function (checkTenderErrors) {
  *
  * const Aerospike = require('aerospike')
  * const op = Aerospike.operations
- *
- * var key = new Aerospike.Key('test', 'demo', 'mykey1')
+ * const key = new Aerospike.Key('test', 'demo', 'mykey1')
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}), 
+ *   }
+ * }
  * var ops = [
  *   op.append('a', 'xyz'),
  *   op.incr('b', 10),
  *   op.read('b')
  * ]
  *
- * Aerospike.connect(function (error, client) {
- *   client.operate(key, ops, function (error, record) {
- *     if (error) {
- *       // handle failure
- *     } else {
- *       console.log('b', record.bins['b']) // value of 'b' returned by the `read` operation
- *     }
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   client.put(key, { a: 'abc', b: 42 }, (error) => {
+ *     if (error) throw error
+ *     client.operate(key, ops, (error, record) => {
+ *       if (error) throw error
+ *       console.log(record.bins) // => { b: 52 }
+ *       client.close()
+ *     })
  *   })
- *   client.close()
  * })
+ * 
  */
 Client.prototype.operate = function (key, operations, metadata, policy, callback) {
   if (typeof policy === 'function') {
@@ -1601,21 +1857,36 @@ Client.prototype.incr = Client.prototype.add
  * @param {writeCallback} [callback] - The function to call when the operation completes with the result of the operation.
  *
  * @example
- *
+ * 
+ * const Aerospike = require('aerospike')
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}), 
+ *   }
+ * }
+ * 
  * const Key = Aerospike.Key
  *
  * var key = new Key('test', 'demo', 'key1')
- * var rec = {
+ * var bins = {
  *   a: 'xyz',
  *   b: 123
  * }
- *
- * client.put(key, rec, function (error, key) {
- *   if (error) {
- *     // handle failure
- *   } else {
- *     // handle success
- *   }
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   client.put(key, bins, (error) => {
+ *     if (error) throw error
+ *     client.get(key, (error, record) => {
+ *       if (error) throw error
+ *       console.log(record)
+ *       client.close()
+ *     })
+ *   })
  * })
  */
 Client.prototype.put = function (key, bins, meta, policy, callback) {
@@ -1672,14 +1943,35 @@ Client.prototype.query = function (ns, set, options) {
  *
  * @example
  *
- * const Key = Aerospike.Key
-
- * client.remove(new Key('test', 'demo', 'key1'), function (error, key) {
- *   if (error) {
- *     // handle failure
- *   } else {
- *     // handle success
+ * const Aerospike = require('aerospike')
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     remove : new Aerospike.RemovePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}), 
  *   }
+ * }
+ * 
+ * const Key = Aerospike.Key
+ *
+ * var key = new Key('test', 'demo', 'key1')
+ * var bins = {
+ *   a: 'xyz',
+ *   b: 123
+ * }
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   client.put(key, bins, (error) => {
+ *     if (error) throw error
+ *     client.remove(key, (error) => {
+ *       if (error) throw error
+ *       console.log("Record removed")
+ *       client.close()
+ *     })
+ *   })
  * })
  */
 Client.prototype.remove = function (key, policy, callback) {
@@ -1729,15 +2021,39 @@ Client.prototype.scan = function (ns, set, options) {
  *
  * @example
  *
- * var key = new Key('test', 'demo', 'key1')
- *
- * client.select(key, ['name', 'age'], (error, record) => {
- *   if (error) {
- *     // handle failure
- *   } else {
- *     // handle success
+ * const Aerospike = require('aerospike')
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}), 
  *   }
+ * }
+ * 
+ * const Key = Aerospike.Key
+ *
+ * var key = new Key('test', 'demo', 'key1')
+ * 
+ * var bins = {
+ *   a: 'xyz',
+ *   b: 123
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   client.put(key, bins, (error) => {
+ *     if (error) throw error
+ *     client.select(key, ['a', 'b'], (error, record) => {
+ *       if (error) throw error
+ *       console.log(record)
+ *       client.close()
+ *     })
+ *   })
  * })
+ *
  */
 Client.prototype.select = function (key, bins, policy, callback) {
   if (typeof policy === 'function') {
@@ -1863,7 +2179,16 @@ Client.prototype.udfRegister = function (udfPath, udfType, policy, callback) {
  * @since v3.8.0
  *
  * @example
- * require('aerospike').connect().then(client => {
+ * 
+ * const Aerospike = require('aerospike')
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
  *   const stats = client.stats()
  *   console.info(stats) // => { commands: { inFlight: 0, queued: 0 },
  *                       //      nodes:
@@ -1875,6 +2200,7 @@ Client.prototype.udfRegister = function (udfPath, udfType, policy, callback) {
  *                       //           asyncConnections: { inPool: 0, inUse: 0 } } ] }
  *   client.close()
  * })
+ * 
  */
 Client.prototype.stats = function () {
   return this.as_client.getStats()
