+++ lib/config.js
@@ -20,6 +20,8 @@ const policy = require('./policy')
 
 const inspect = Symbol.for('nodejs.util.inspect.custom')
 
+
+class Config {
 /**
  * @class Config
  * @classdesc config command.
@@ -33,14 +35,14 @@ const inspect = Symbol.for('nodejs.util.inspect.custom')
  * @example
  *
  * const Aerospike = require('aerospike')
- *
+ * 
  * let config = {
  *   hosts: '192.168.1.10,192.168.1.11',
  *   user: process.env.DATABASE_USER,
  *   password: process.env.DATABASE_PASSWORD,
  *   policies: {
  *     read: new Aerospike.ReadPolicy({
- *       totalTimeout: 500
+ *       totalTimeout: 0
  *     })
  *   },
  *   log: {
@@ -57,13 +59,12 @@ const inspect = Symbol.for('nodejs.util.inspect.custom')
  *   .catch(error => {
  *     console.error('Failed to connect to cluster: %s', error.message)
  *   })
+ *
+ *
+ * // Initializes a new client configuration from the given config values.
+ *
+ * @param {Object} [config] configuration values
  */
-class Config {
-  /**
-   * Initializes a new client configuration from the given config values.
-   *
-   * @param {Object} [config] configuration values
-   */
   constructor (config) {
     config = config || {}
 
@@ -259,13 +260,18 @@ class Config {
      *
      * const Aerospike = require('aerospike')
      *
-     * let config = {
+     * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+     * var config = {
+     *   hosts: '192.168.33.10:3000',
      *   policies: {
      *     write: new Aerospike.WritePolicy({
-     *       key: Aerospike.policy.key.SEND
+     *       key: Aerospike.policy.key.SEND,
+     *       socketTimeout : 0,
+     *       totalTimeout : 0
      *     })
      *   }
      * }
+     * 
      * let key = new Aerospike.Key('test', 'demo', 123)
      *
      * Aerospike.connect(config)
@@ -273,8 +279,8 @@ class Config {
      *     return client.put(key, {int: 42})
      *       .then(() => client.close())
      *       .catch(error => {
-     *         client.close()
      *         throw error
+     *         client.close()
      *       })
      *   })
      *   .catch(console.error)
@@ -298,10 +304,13 @@ class Config {
      * @example <caption>Enabling debug logging to a separate log file</caption>
      *
      * const Aerospike = require('aerospike')
+     * 
      * const fs = require('fs')
      *
-     * var debuglog = fs.openSync('./debug.log')
+     * var debuglog = fs.openSync('./debug.log', 'w')
+     * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
      * var config = {
+     *   hosts: '192.168.33.10:3000',
      *   log: {
      *     level: Aerospike.log.DEBUG,
      *     file: debuglog
@@ -309,7 +318,7 @@ class Config {
      * }
      * Aerospike.connect(config, (err, client) => {
      *   if (err) throw err
-     *   // ...
+     *   console.log("Connected. Now closing connection.")
      *   client.close()
      * })
      */
