+++ lib/double.js
@@ -32,30 +32,34 @@
  *
  * const Aerospike = require('aerospike')
  * const Double = Aerospike.Double
- * const client = Aerospike.client().connect((error) => {
- *   if (error) throw error
- * })
- * const key = new Aerospike.Key('test', 'demo', 'myDouble')
- *
- * var record = { d: 3.1415 }
- * client.put(key, record, (error) => {
- *   if (error) throw error
- * })
- *
- * function incr (value) {
- *   // wrap value in Double since we can't be sure it would be converted to
- *   // double automatically, e.g. 1.0
- *   client.operate(key, [Aerospike.operations.incr('d', new Double(value))], (error) => {
- *     if (error) throw error
- *   })
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
  * }
- *
- * incr(6.283)
- * incr(1.0)
- *
- * client.get(key, (error, record) => {
- *   console.log(record)
- *   client.close()
+ * // (1.0) must be wrapped with "Double" in order to be added to another double.
+ * // (6.283) does not need to be wrapped, but it may be wrapped if convenient.
+ * ops = [Aerospike.operations.incr('d', 6.283),
+ *        Aerospike.operations.incr('d', new Double(1.0))]
+ * const key = new Aerospike.Key('test', 'demo', 'myDouble')
+ * var record = { d: 3.1415 }
+ * Aerospike.connect(config, (error, client) => {
+ *   if (error) throw error
+ *   client.put(key, record, (error) => {
+ *     if (error) throw error
+ *     client.operate(key, ops, (error) => {
+ *       if (error) throw error
+ *       client.get(key, (error, record) => {
+ *         console.log(record.bins.d) // => 10.4245
+ *         client.close()
+ *       })
+ *     })
+ *   })
  * })
  */
 function Double (value) {
