+++ lib/error.js
@@ -28,6 +28,17 @@ const status = require('./status')
  * @example <caption>Expected output: "Error: 127.0.0.1:3000 Record does not exist in database. May be returned by read, or write with policy Aerospike.policy.exists.UPDATE [2]"</caption>
  *
  * const Aerospike = require('aerospike')
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * 
  * let key = new Aerospike.Key('test', 'key', 'does_not_exist')
  * Aerospike.connect()
  *   .then(client => {
@@ -152,8 +163,18 @@ class AerospikeError extends Error {
    * @example <caption>Closing the client connection, when an error occurs:</caption>
    *
    * const Aerospike = require('aerospike')
-   *
-   * Aerospike.connect().then(async client => {
+   * 
+   * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+   * var config = {
+   *   hosts: '192.168.33.10:3000',
+   *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+   *   policies: {
+   *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+   *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0})
+   *   }
+   * }
+   * 
+   * Aerospike.connect(config).then(async client => {
    *   await client.put(new Aerospike.Key('demo', 'test', 'foo'), { 'foo': 'bar' })
    *   client.close()
    * }).catch(error => {
