+++ lib/exp.js
@@ -31,11 +31,11 @@ const BIN_TYPE_UNDEF = 0
  * @summary {@link module:aerospike/exp|aerospike/exp} module
  *
  * @description This module defines a filter expression that is
- * a mechanism for additional filtering.
- * The resultset of a secondary index (SI) query or primary index (PI)
- * query (scan) operation can be filtered through the QueryPolicy
- * and ScanPolicy classes.
- * It can also filter single key operations and
+ * a mechanism for additional filtering. 
+ * 
+ * The resultset of a primary index (PI) query (scan) or secondary
+ * index (SI) query operation can be filtered through the QueryPolicy
+ * and ScanPolicy classes. It can also filter single key operations and
  * batch operations through the filterExpression field of their policy class.
  *
  * Filter Expressions replace PredicateExpression filtering,
@@ -51,54 +51,86 @@ const BIN_TYPE_UNDEF = 0
  * @see {@link QueryPolicy#filterExpression}
  *
  * @example
- *
+ * 
+ * <caption>Expressions using the operate API</caption>
  * const Aerospike = require('aerospike')
  * const op = Aerospike.operations
  * const exp = Aerospike.exp
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
  * const tempBin = 'ExpVar' // this bin is to hold expression read operation output
  *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROSPIKE SERVER NODE HERE!
+ * const config = {  
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ *   
  * var ops = [
  *   op.append('a', 'xyz'),
  *   op.incr('b', 10),
  *   exp.operations.read(tempBin,
  *           exp.add(exp.binInt('b'), exp.binInt('b')),
  *          0),
+ *   op.read('a'),
  *   op.read('b')
  * ]
  *
- * //using operate API
- *
- * Aerospike.connect((error, client) => {
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *   client.put(key, { a: 'abc', b: 42 }, (error) => {
  *     if (error) throw error
  *     client.operate(key, ops, (error, record) => {
  *       if (error) throw error
- *       console.log(record) // => { b: 52, ExpVar: 104 }
+ *       console.log(record.bins) // => { a: 'abcxyz', b: 52, ExpVar: 104 }
  *       client.close()
  *     })
  *   })
  * })
+ * 
+ * @example
+ *  
+ * <caption>Expressions using the query API</caption>
+ * const Aerospike = require('aerospike')
+ * const op = Aerospike.operations
+ * const exp = Aerospike.exp
+ * const key = new Aerospike.Key('test', 'demo', 'mykey1')
+ * const tempBin = 'ExpVar' // this bin is to hold expression read operation output
  *
- * //using query API
- *
- *  Aerospike.connect((error, client) => {
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROSPIKE SERVER NODE HERE!
+ * const config = {  
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     query : new Aerospike.QueryPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0, exists : Aerospike.policy.exists.REPLACE})
+ *   }
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *   client.put(key, { a: 'abc', b: 42 }, (error) => {
  *     if (error) throw error
+ *     var query = client.query('test', 'demo')
  *     const queryPolicy = { filterExpression:  exp.eq(exp.binInt('b'), exp.int(42))}
- *     query.nobins = true
- *     const stream = query.foreach(queryPolicy) => {
- *       if (error) throw error
- *       // expect(record.bins).to.be.empty()
- *       // expect(record.gen).to.be.ok()
- *       // expect(record.ttl).to.be.ok()
- *       console.log(record)
+ *     query.nobins = false
+ *     const stream = query.foreach(queryPolicy)
+ *     stream.on('error', (error) => {
+ *       console.error(error)
+ *       throw error
+ *     })
+ *     stream.on('data', (record) => {
+ *       console.info(record.bins) // => { a: 'abc', b: 42} 
+ *     })
+ *     stream.on('end', () => {
  *       client.close()
  *     })
  *   })
  * })
+ * 
  */
 
 /*********************************************************************************
@@ -849,1891 +881,34 @@ exports.var = (varName) => [
   ..._rawStr(varName)
 ]
 
-/*********************************************************************************
- * LIST READ EXPRESSIONS
- *********************************************************************************/
-
-function getListType (type, returnType, isMulti) {
-  let expected = type
-
-  switch (returnType & ~lists.returnType.INVERTED) {
-    case lists.returnType.INDEX:
-    case lists.returnType.REVERSE_INDEX:
-    case lists.returnType.RANK:
-    case lists.returnType.REVERSE_RANK:
-      expected = isMulti ? exp.type.LIST : exp.type.INT
-      break
-    case lists.returnType.COUNT:
-      expected = exp.type.INT
-      break
-    case lists.returnType.VALUE:
-      if (isMulti) {
-        expected = exp.type.LIST
-      }
-      break
-    case lists.returnType.NONE:
-    default:
-      throw new TypeError('either set the value type as auto or match with return object data type')
-  }
-
-  if (type === exp.type.AUTO || type === expected) {
-    return expected
-  }
-
-  throw new TypeError('either set the value type as auto or match with return object data type')
-}
-
-const _rtype = _valueExp(exp.ops.VAL_RTYPE, 'intVal')
-
-const _listRead = (type, returnType, isMulti) => [
-  { op: exp.ops.CALL, count: 5 },
-  ..._rtype(getListType(type, returnType, isMulti)),
-  ..._int(exp.sys.CALL_CDT)
-]
-
-const _listModify = (ctx, policy, op, param, extraParam) => [
-  { op: exp.ops.CALL, count: 5 },
-  ..._rtype(cdtCtx.getContextType(ctx, exp.type.LIST)),
-  ..._int(exp.sys.CALL_CDT | exp.sys.FLAG_MODIFY_LOCAL),
-  { op: exp.ops.CALL_VOP_START, count: 1 + param + (policy ? extraParam : 0), ctx },
-  ..._int(op)
-]
-
-const _listStart = (ctx, op, param) => [
-  { op: exp.ops.CALL_VOP_START, count: 1 + param, ctx },
-  ..._int(op)
-]
-
-exports.lists = {
 /**
- * Create expression that returns list size.
- *
- * @param {Object} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (integer expression)
- */
-  size: (bin, ctx = null) => [
-    ..._listRead(exp.type.AUTO, lists.returnType.COUNT, false),
-    ..._listStart(ctx, lists.opcodes.SIZE, 0),
-    ...bin
-  ],
-
-  /**
- * Create expression that selects list items identified by value and returns selected
- * data specified by returnType.
- *
- * @param {Object} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByValue: (bin, value, returnType, ctx = null) => [
-    ..._listRead(exp.type.AUTO, returnType, true),
-    ..._listStart(ctx, lists.opcodes.GET_ALL_BY_VALUE, 2),
-    ..._int(returnType),
-    ...value,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects list items identified by value range and returns selected
- * data specified by returnType.
- *
- * @param {Object} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} begin Begin value expression.
- * @param {AerospikeExp} end End value expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByValueRange: (bin, begin, end, returnType, ctx = null) => [
-    ..._listRead(exp.type.AUTO, returnType, true),
-    ..._listStart(ctx, lists.opcodes.GET_BY_VALUE_INTERVAL, 3),
-    ..._int(returnType),
-    ...begin,
-    ...end,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects list items identified by values and returns selected
- * data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} value Values list expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByValueList: (bin, value, returnType, ctx = null) => [
-    ..._listRead(exp.type.AUTO, returnType, true),
-    ..._listStart(ctx, lists.opcodes.GET_BY_VALUE_LIST, 2),
-    ..._int(returnType),
-    ...value,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects list items nearest to value and greater by relative rank
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} value Values list expression.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByRelRankRangeToEnd: (bin, value, rank, returnType, ctx = null) => [
-    ..._listRead(exp.type.AUTO, returnType, true),
-    ..._listStart(ctx, lists.opcodes.GET_BY_VALUE_REL_RANK_RANGE, 3),
-    ..._int(returnType),
-    ...value,
-    ...rank,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects list items nearest to value and greater by relative rank with a
- * count limit and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} value Values list expression.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByRelRankRange: (bin, value, rank, count, returnType, ctx = null) => [
-    ..._listRead(exp.type.AUTO, returnType, true),
-    ..._listStart(ctx, lists.opcodes.GET_BY_VALUE_REL_RANK_RANGE, 4),
-    ..._int(returnType),
-    ...value,
-    ...rank,
-    ...count,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects list item identified by index
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} valueType expression value type.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (valueType expression)
- */
-  getByIndex: (bin, index, valueType, returnType, ctx = null) => [
-    ..._listRead(valueType, returnType, false),
-    ..._listStart(ctx, lists.opcodes.GET_BY_INDEX, 2),
-    ..._int(returnType),
-    ...index,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects list items starting at specified index to the end of list
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByIndexRangeToEnd: (bin, index, returnType, ctx = null) => [
-    ..._listRead(exp.type.AUTO, returnType, true),
-    ..._listStart(ctx, lists.opcodes.GET_BY_INDEX_RANGE, 2),
-    ..._int(returnType),
-    ...index,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects "count" list items starting at specified index
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByIndexRange: (bin, index, count, returnType, ctx = null) => [
-    ..._listRead(exp.type.AUTO, returnType, true),
-    ..._listStart(ctx, lists.opcodes.GET_BY_INDEX_RANGE, 3),
-    ..._int(returnType),
-    ...index,
-    ...count,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects list item identified by rank
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} valueType expression value type.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (valueType expression)
- */
-  getByRank: (bin, rank, valueType, returnType, ctx = null) => [
-    ..._listRead(valueType, returnType, false),
-    ..._listStart(ctx, lists.opcodes.GET_BY_RANK, 2),
-    ..._int(returnType),
-    ...rank,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects list items starting at specified rank to the last ranked item
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByRankRangeToEnd: (bin, rank, returnType, ctx = null) => [
-    ..._listRead(exp.type.AUTO, returnType, true),
-    ..._listStart(ctx, lists.opcodes.GET_BY_RANK_RANGE, 2),
-    ..._int(returnType),
-    ...rank,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects "count" list items starting at specified rank
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByRankRange: (bin, rank, count, returnType, ctx = null) => [
-    ..._listRead(exp.type.AUTO, returnType, true),
-    ..._listStart(ctx, lists.opcodes.GET_BY_RANK_RANGE, 3),
-    ..._int(returnType),
-    ...rank,
-    ...count,
-    ...bin
-  ],
-
-  /*********************************************************************************
- * LIST MODIFY EXPRESSIONS
- *********************************************************************************/
-
-  /**
- * Create expression that appends value to end of list.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {Object} policy Optional list write policy.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  append: (bin, value, policy = null, ctx = null) => [
-    ..._listModify(ctx, policy, lists.opcodes.APPEND, 1, 2),
-    ...value,
-    { op: exp.ops.CDT_LIST_CRMOD, listPolicy: policy },
-    ...bin
-  ],
-
-  /**
- * Create expression that appends list items to end of list.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {Object} policy Optional list write policy.
- * @param {AerospikeExp} value List items expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  appendItems: (bin, value, policy = null, ctx = null) => [
-    ..._listModify(ctx, policy, lists.opcodes.APPEND_ITEMS, 1, 2),
-    ...value,
-    { op: exp.ops.CDT_LIST_CRMOD, listPolicy: policy },
-    ...bin
-  ],
-
-  /**
- * Create expression that inserts value to specified index of list.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {Object} policy Optional list write policy.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  insert: (bin, value, idx, policy = null, ctx = null) => [
-    ..._listModify(ctx, policy, lists.opcodes.INSERT, 2, 1),
-    ...idx,
-    ...value,
-    { op: exp.ops.CDT_LIST_CRMOD, listPolicy: policy },
-    ...bin
-  ],
-
-  /**
- * Create expression that inserts each input list item starting at specified index of list.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {Object} policy Optional list write policy.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} value List items expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  insertItems: (bin, value, idx, policy = null, ctx = null) => [
-    ..._listModify(ctx, policy, lists.opcodes.INSERT_ITEMS, 2, 1),
-    ...idx,
-    ...value,
-    { op: exp.ops.CDT_LIST_CRMOD, listPolicy: policy },
-    ...bin
-  ],
-
-  /**
- * Create expression that increments list[index] by value.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {Object} policy Optional list write policy.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  increment: (bin, value, idx, policy = null, ctx = null) => [
-    ..._listModify(ctx, policy, lists.opcodes.INCREMENT, 2, 2),
-    ...idx,
-    ...value,
-    { op: exp.ops.CDT_LIST_CRMOD, listPolicy: policy },
-    ...bin
-  ],
-
-  /**
- * Create expression that sets item value at specified index in list.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {Object} policy Optional list write policy.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  set: (bin, value, idx, policy = null, ctx = null) => [
-    ..._listModify(ctx, policy, lists.opcodes.SET, 2, 1),
-    ...idx,
-    ...value,
-    { op: exp.ops.CDT_LIST_CRMOD, listPolicy: policy },
-    ...bin
-  ],
-
-  /**
- * Create expression that removes all items in list.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  clear: (bin, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.CLEAR, 0, 0),
-    ...bin
-  ],
-
-  /**
- * Create expression that sorts list.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {number} order Sort order flags.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  sort: (bin, order, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.SORT, 1, 0),
-    ..._int(order),
-    ...bin
-  ],
-
-  /**
- * Create expression that removes list items identified by value.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  removeByValue: (bin, value, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.REMOVE_ALL_BY_VALUE, 2, 0),
-    ..._int(lists.RETURN_NONE),
-    ...value,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes list items identified by values.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} values Values list expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  removeByValueList: (bin, values, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.REMOVE_BY_VALUE_LIST, 2, 0),
-    ..._int(lists.RETURN_NONE),
-    ...values,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes list items identified by value range
- * (begin inclusive, end exclusive). If begin is nil, the range is less than end.
- * If end is infinity, the range is greater than equal to begin.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} begin Begin value expression.
- * @param {AerospikeExp} end End value expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  removeByValueRange: (bin, end, begin, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.REMOVE_BY_VALUE_INTERVAL, 3, 0),
-    ..._int(lists.RETURN_NONE),
-    ...begin,
-    ...end,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes list items nearest to value and greater by relative rank.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  removeByRelRankRangeToEnd: (bin, rank, value, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.REMOVE_BY_VALUE_REL_RANK_RANGE, 3, 0),
-    ..._int(lists.RETURN_NONE),
-    ...value,
-    ...rank,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes list items nearest to value and greater by relative rank with a
- * count limit.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  removeByRelRankRange: (bin, count, rank, value, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.REMOVE_BY_VALUE_REL_RANK_RANGE, 4, 0),
-    ..._int(lists.RETURN_NONE),
-    ...value,
-    ...rank,
-    ...count,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes list item identified by index.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  removeByIndex: (bin, idx, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.REMOVE_BY_INDEX, 2, 0),
-    ..._int(lists.RETURN_NONE),
-    ...idx,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes list items starting at specified index to the end of list.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  removeByIndexRangeToEnd: (bin, idx, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.REMOVE_BY_INDEX_RANGE, 2, 0),
-    ..._int(lists.RETURN_NONE),
-    ...idx,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes "count" list items starting at specified index.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  removeByIndexRange: (bin, count, idx, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.REMOVE_BY_INDEX_RANGE, 3, 0),
-    ..._int(lists.RETURN_NONE),
-    ...idx,
-    ...count,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes list item identified by rank.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  removeByRank: (bin, rank, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.REMOVE_BY_RANK, 2, 0),
-    ..._int(lists.RETURN_NONE),
-    ...rank,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes list items starting at specified rank to the last ranked item.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  removeByRankRangeToEnd: (bin, rank, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.REMOVE_BY_RANK_RANGE, 2, 0),
-    ..._int(lists.RETURN_NONE),
-    ...rank,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes "count" list items starting at specified rank.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin List bin or list value expression.
- * @return {AerospikeExp} (list expression)
- */
-  removeByRankRange: (bin, count, rank, ctx = null) => [
-    ..._listModify(ctx, null, lists.opcodes.REMOVE_BY_RANK_RANGE, 3, 0),
-    ..._int(lists.RETURN_NONE),
-    ...rank,
-    ...count,
-    ...bin
-  ]
-}
-
-/*********************************************************************************
- * MAP MODIFY EXPRESSIONS
- *********************************************************************************/
-
-function getMapType (type, returnType, isMulti) {
-  let expected = type
-
-  switch (returnType & ~maps.returnType.INVERTED) {
-    case maps.returnType.INDEX:
-    case maps.returnType.REVERSE_INDEX:
-    case maps.returnType.RANK:
-    case maps.returnType.REVERSE_RANK:
-      expected = isMulti ? exp.type.LIST : exp.type.INT
-      break
-    case maps.returnType.COUNT:
-      expected = exp.type.INT
-      break
-    case maps.returnType.KEY:
-    case maps.returnType.VALUE:
-      if (isMulti) {
-        expected = exp.type.LIST
-      }
-      break
-    case maps.returnType.KEY_VALUE:
-      expected = exp.type.MAP
-      break
-    case maps.returnType.NONE:
-    default:
-      throw new TypeError('either set the value type as auto or match with return object data type')
-  }
-
-  if (type === exp.type.AUTO || type === expected) {
-    return expected
-  }
-
-  throw new TypeError('either set the value type as auto or match with return object data type')
-}
-
-const _mapRead = (type, returnType, isMulti) => [
-  { op: exp.ops.CALL, count: 5 },
-  ..._rtype(getMapType(type, returnType, isMulti)),
-  ..._int(exp.sys.CALL_CDT)
-]
-
-const _mapStart = (ctx, op, param) => [
-  { op: exp.ops.CALL_VOP_START, count: 1 + param, ctx },
-  ..._int(op)
-]
-
-const _mapModify = (ctx, policy, op, param, extraParam) => [
-  { op: exp.ops.CALL, count: 5 },
-  ..._rtype(cdtCtx.getContextType(ctx, exp.type.MAP)),
-  ..._int(exp.sys.CALL_CDT | exp.sys.FLAG_MODIFY_LOCAL),
-  { op: exp.ops.CALL_VOP_START, count: 1 + param + (policy ? extraParam : 0), ctx },
-  ..._int(op)
-]
-
-exports.maps = {
-
-  /**
- * Create expression that writes key/val item to map bin.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {Object} policy Optional map write policy.
- * @param {AerospikeExp} key Key expression.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  put: (bin, value, key, policy = null, ctx = null) => [
-    ..._mapModify(ctx, policy, maps.opcodes.PUT, 2, 2),
-    ...key,
-    ...value,
-    { op: exp.ops.CDT_MAP_CRMOD, mapPolicy: policy },
-    ...bin
-  ],
-
-  /**
- * Create expression that writes each map item to map bin.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {Object} policy Optional map write policy.
- * @param {AerospikeExp} map Source map expression.
- * @param {AerospikeExp} bin Target map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  putItems: (bin, map, policy = null, ctx = null) => [
-    ..._mapModify(ctx, policy, maps.opcodes.PUT_ITEMS, 1, 2),
-    ...map,
-    { op: exp.ops.CDT_MAP_CRMOD, mapPolicy: policy },
-    ...bin
-  ],
-
-  /**
- * Create expression that increments values by incr for all items identified by key.
- * Valid only for numbers.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {Object} policy Optional map write policy.
- * @param {AerospikeExp} key Key expression.
- * @param {AerospikeExp} value Increment value number expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  increment: (bin, value, key, policy = null, ctx = null) => [
-    ..._mapModify(ctx, policy, maps.opcodes.INCREMENT, 2, 1),
-    ...key,
-    ...value,
-    { op: exp.ops.CDT_MAP_CRMOD, mapPolicy: policy },
-    ...bin
-  ],
-
-  /**
- * Create expression that removes all items in map.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  clear: (bin, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.CLEAR, 0, 0),
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map item identified by key.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} key Key expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByKey: (bin, key, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_KEY, 2, 0),
-    ..._int(maps.RETURN_NONE),
-    ...key,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map items identified by keys.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} keys List expression of keys to remove.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByKeyList: (bin, keys, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_KEY_LIST, 2, 0),
-    ..._int(maps.RETURN_NONE),
-    ...keys,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map items identified by key range
- * (begin inclusive, end exclusive). If begin is nil, the range is less than end.
- * If end is infinity, the range is greater than equal to begin.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} begin Begin value expression.
- * @param {AerospikeExp} end End value expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByKeyRange: (bin, end, begin, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_KEY_INTERVAL, 3, 0),
-    ..._int(maps.RETURN_NONE),
-    ...begin,
-    ...end,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map items nearest to key and greater by index.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} key Key expression.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByKeyRelIndexRangeToEnd: (bin, idx, key, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_KEY_REL_INDEX_RANGE, 3, 0),
-    ..._int(maps.RETURN_NONE),
-    ...key,
-    ...idx,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map items nearest to key and greater by index with a count limit.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} key Key expression.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByKeyRelIndexRange: (bin, count, idx, key, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_KEY_REL_INDEX_RANGE, 4, 0),
-    ..._int(maps.RETURN_NONE),
-    ...key,
-    ...idx,
-    ...count,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map items identified by value.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByValue: (bin, value, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_ALL_BY_VALUE, 2, 0),
-    ..._int(maps.RETURN_NONE),
-    ...value,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map items identified by values.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} values Values list expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByValueList: (bin, values, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_VALUE_LIST, 2, 0),
-    ..._int(maps.RETURN_NONE),
-    ...values,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map items identified by value range
- * (begin inclusive, end exclusive). If begin is nil, the range is less than end.
- * If end is infinity, the range is greater than equal to begin.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} begin Begin value expression.
- * @param {AerospikeExp} end End value expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByValueRange: (bin, end, begin, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_VALUE_INTERVAL, 3, 0),
-    ..._int(maps.RETURN_NONE),
-    ...begin,
-    ...end,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map items nearest to value and greater by relative rank.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByValueRelRankRangeToEnd: (bin, rank, value, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_VALUE_REL_RANK_RANGE, 3, 0),
-    ..._int(maps.RETURN_NONE),
-    ...value,
-    ...rank,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map items nearest to value and greater by relative rank with a
- * count limit.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByValueRelRankRange: (bin, count, rank, value, key, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_VALUE_REL_RANK_RANGE, 4, 0),
-    ..._int(maps.RETURN_NONE),
-    ...value,
-    ...rank,
-    ...count,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map item identified by index.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByIndex: (bin, idx, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_INDEX, 2, 0),
-    ..._int(maps.RETURN_NONE),
-    ...idx,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map items starting at specified index to the end of map.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByIndexRangeToEnd: (bin, idx, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_INDEX_RANGE, 2, 0),
-    ..._int(maps.RETURN_NONE),
-    ...idx,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes "count" map items starting at specified index.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByIndexRange: (bin, count, idx, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.PUT, 3, 0),
-    ..._int(maps.RETURN_NONE),
-    ...idx,
-    ...count,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map item identified by rank.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByRank: (bin, rank, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_RANK, 2, 0),
-    ..._int(maps.RETURN_NONE),
-    ...rank,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes map items starting at specified rank to the last ranked item.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByRankRangeToEnd: (bin, rank, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_RANK_RANGE, 2, 0),
-    ..._int(maps.RETURN_NONE),
-    ...rank,
-    ...bin
-  ],
-
-  /**
- * Create expression that removes "count" map items starting at specified rank.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (map expression)
- */
-  removeByRankRange: (bin, count, rank, ctx = null) => [
-    ..._mapModify(ctx, null, maps.opcodes.REMOVE_BY_RANK_RANGE, 3, 0),
-    ..._int(maps.RETURN_NONE),
-    ...rank,
-    ...count,
-    ...bin
-  ],
-
-  /*********************************************************************************
- * MAP READ EXPRESSIONS
- *********************************************************************************/
-
-  /**
- * Create expression that returns map size.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (integer expression)
- */
-  size: (bin, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, maps.returnType.COUNT, false),
-    ..._mapStart(ctx, maps.opcodes.SIZE, 0),
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map item identified by key
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} valueType expression value type.
- * @param {AerospikeExp} key Key expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByKey: (bin, key, valueType, returnType, ctx = null) => [
-    ..._mapRead(valueType, returnType, false),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_KEY, 2),
-    ..._int(returnType),
-    ...key,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map items identified by key range
- * (begin inclusive, end exclusive). If begin is nil, the range is less than end.
- * If end is infinity, the range is greater than equal to begin.
- * Expression returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} begin Begin key expression.
- * @param {AerospikeExp} end End key expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByKeyRange: (bin, end, begin, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_KEY_INTERVAL, 3),
-    ..._int(returnType),
-    ...begin,
-    ...end,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map items identified by keys
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} keys Keys list expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByKeyList: (bin, keys, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_KEY_INTERVAL, 2),
-    ..._int(returnType),
-    ...keys,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map items nearest to key and greater by index
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} key Key expression.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByKeyRelIndexRangeToEnd: (bin, idx, key, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_KEY_INTERVAL, 3),
-    ..._int(returnType),
-    ...key,
-    ...idx,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map items nearest to key and greater by index with a count limit.
- * Expression returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} key Key expression.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByKeyRelIndexRange: (bin, count, idx, key, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_KEY_REL_INDEX_RANGE, 4),
-    ..._int(returnType),
-    ...key,
-    ...idx,
-    ...count,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map items identified by value
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByValue: (bin, value, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_ALL_BY_VALUE, 2),
-    ..._int(returnType),
-    ...value,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map items identified by value range
- * (begin inclusive, end exclusive). If begin is nil, the range is less than end.
- * If end is infinity, the range is greater than equal to begin.
- * Expression returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} begin Begin value expression.
- * @param {AerospikeExp} end End value expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByValueRange: (bin, end, begin, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_VALUE_INTERVAL, 3),
-    ..._int(returnType),
-    ...begin,
-    ...end,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map items identified by values
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} values Values list expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByValueList: (bin, values, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_VALUE_INTERVAL, 2),
-    ..._int(returnType),
-    ...values,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map items nearest to value and greater by relative rank.
- * Expression returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByValueRelRankRangeToEnd: (bin, rank, value, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_VALUE_REL_RANK_RANGE, 3),
-    ..._int(returnType),
-    ...value,
-    ...rank,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map items nearest to value and greater by relative rank with a
- * count limit. Expression returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} value Value expression.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByValueRelRankRange: (bin, count, rank, value, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_VALUE_REL_RANK_RANGE, 4),
-    ..._int(returnType),
-    ...value,
-    ...rank,
-    ...count,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map item identified by index
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} valueType expression value type.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByIndex: (bin, idx, valueType, returnType, ctx = null) => [
-    ..._mapRead(valueType, returnType, false),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_INDEX, 2),
-    ..._int(returnType),
-    ...idx,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map items starting at specified index to the end of map
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByIndexRangeToEnd: (bin, idx, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_INDEX_RANGE, 2),
-    ..._int(returnType),
-    ...idx,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects "count" map items starting at specified index
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} idx Index integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByIndexRange: (bin, count, idx, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_INDEX_RANGE, 3),
-    ..._int(returnType),
-    ...idx,
-    ...count,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map item identified by rank
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} valueType expression value type.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByRank: (bin, rank, valueType, returnType, ctx = null) => [
-    ..._mapRead(valueType, returnType, false),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_RANK, 2),
-    ..._int(returnType),
-    ...rank,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects map items starting at specified rank to the last ranked item
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByRankRangeToEnd: (bin, rank, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_RANK_RANGE, 2),
-    ..._int(returnType),
-    ...rank,
-    ...bin
-  ],
-
-  /**
- * Create expression that selects "count" map items starting at specified rank
- * and returns selected data specified by returnType.
- *
- * @param {AerospikeExp} ctx Optional context path for nested CDT.
- * @param {AerospikeExp} returnType Return type.
- * @param {AerospikeExp} rank Rank integer expression.
- * @param {AerospikeExp} count Count integer expression.
- * @param {AerospikeExp} bin Map bin or map value expression.
- * @return {AerospikeExp} (expression)
- */
-  getByRankRange: (bin, count, rank, returnType, ctx = null) => [
-    ..._mapRead(exp.type.AUTO, returnType, true),
-    ..._mapStart(ctx, maps.opcodes.GET_BY_RANK_RANGE, 3),
-    ..._int(returnType),
-    ...rank,
-    ...count,
-    ...bin
-  ]
-}
-
-/*********************************************************************************
- * BIT MODIFY EXPRESSIONS
- *********************************************************************************/
-
-const _bitModify = () => [
-  { op: exp.ops.CALL, count: 5 },
-  ..._rtype(exp.type.BLOB),
-  ..._int(exp.sys.CALL_BITS | exp.sys.FLAG_MODIFY_LOCAL)
-]
-
-const _bitModStart = (op, param) => [
-  ..._bitModify(),
-  { op: exp.ops.CALL_VOP_START, count: 1 + param },
-  ..._int(op)
-]
-
-const _bitRead = (returnType) => [
-  { op: exp.ops.CALL, count: 5 },
-  ..._rtype(returnType),
-  ..._int(exp.sys.CALL_BITS)
-]
-
-const _bitReadStart = (returnType, op, param) => [
-  ..._bitRead(returnType),
-  { op: exp.ops.CALL_VOP_START, count: 1 + param },
-  ..._int(op)
-]
-
-exports.bit = {
-
-  /**
- * Create an expression that performs bit resize operation.
- *
- * @param {Object} policy bit policy value.
- * @param {number} byteSize Number of bytes the resulting blob should occupy.
- * @param {number} flags bit resize flags value.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin byteSize bytes.
- */
-  reSize: (bin, flags, byteSize, policy = null) => [
-    ..._bitModStart(bits.opcodes.RESIZE, 3),
-    ...byteSize,
-    ..._uint((policy ? policy.flags : 0)),
-    ..._uint(flags),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit insert operation.
- *
- * @param {Object} policy bit policy value.
- * @param {AerospikeExp} byteOffset Byte index of where to insert the value.
- * @param {AerospikeExp} value Blob expression containing the bytes to insert.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin resulting blob containing the inserted bytes.
- */
-  insert: (bin, value, byteOffset, policy = null) => [
-    ..._bitModStart(bits.opcodes.INSERT, 3),
-    ...byteOffset,
-    ...value,
-    ..._uint((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit remove operation.
- *
- * @param {Object} policy bit policy value.
- * @param {AerospikeExp} byteOffset Byte index of where to remove from.
- * @param {number} byteSize Number of bytes to remove.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin resulting blob with the bytes removed.
- */
-  remove: (bin, byteSize, byteOffset, policy = null) => [
-    ..._bitModStart(bits.opcodes.REMOVE, 3),
-    ...byteOffset,
-    ...byteSize,
-    ..._uint((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit set operation.
- *
- * @param {Object} policy bit policy value.
- * @param {AerospikeExp} bitOffset Bit index of where to start writing.
- * @param {AerospikeExp} bitSize Number of bytes to overwrite.
- * @param {AerospikeExp} value Blob expression containing bytes to write.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin resulting blob with the bytes overwritten.
- */
-  set: (bin, value, bitSize, bitOffset, policy = null) => [
-    ..._bitModStart(bits.opcodes.SET, 4),
-    ...bitOffset,
-    ...bitSize,
-    ...value,
-    ..._uint((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit or operation.
- *
- * @param {Object} policy bit policy value.
- * @param {AerospikeExp} bitOffset Bit index of where to start operation.
- * @param {AerospikeExp} bitSize Number of bytes to be operated on.
- * @param {AerospikeExp} value Blob expression containing bytes to write.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin resulting blob with the bytes operated on.
- */
-  or: (bin, value, bitSize, bitOffset, policy = null) => [
-    ..._bitModStart(bits.opcodes.OR, 4),
-    ...bitOffset,
-    ...bitSize,
-    ...value,
-    ..._uint((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit xor operation.
- *
- * @param {Object} policy bit policy value.
- * @param {AerospikeExp} bitOffset Bit index of where to start operation.
- * @param {AerospikeExp} bitSize Number of bits to be operated on.
- * @param {AerospikeExp} value Blob expression containing bytes to write.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin resulting blob with the bytes operated on.
- */
-  xor: (bin, value, bitSize, bitOffset, policy = null) => [
-    ..._bitModStart(bits.opcodes.XOR, 4),
-    ...bitOffset,
-    ...bitSize,
-    ...value,
-    ..._uint((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit and operation.
- *
- * @param {Object} policy bit policy value.
- * @param {AerospikeExp} bitOffset Bit index of where to start operation.
- * @param {AerospikeExp} bitSize Number of bits to be operated on.
- * @param {AerospikeExp} value Blob expression containing bytes to write.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin resulting blob with the bytes operated on.
- */
-  and: (bin, value, bitSize, bitOffset, policy = null) => [
-    ..._bitModStart(bits.opcodes.AND, 4),
-    ...bitOffset,
-    ...bitSize,
-    ...value,
-    ..._uint((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit not operation.
- *
- * @param {Object} policy bit policy value.
- * @param {AerospikeExp} bitOffset Bit index of where to start operation.
- * @param {AerospikeExp} bitSize Number of bits to be operated on.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin resulting blob with the bytes operated on.
- */
-  not: (bin, bitSize, bitOffset, policy = null) => [
-    ..._bitModStart(bits.opcodes.NOT, 3),
-    ...bitOffset,
-    ...bitSize,
-    ..._uint((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs an bit lshift operation.
- *
- * @param {Object} policy bit policy value.
- * @param {AerospikeExp} bitOffset Bit index of where to start operation.
- * @param {AerospikeExp} bitSize Number of bits to be operated on.
- * @param {number} shift Number of bits to shift by.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin resulting blob with the bytes operated on.
- */
-  lShift: (bin, shift, bitSize, bitOffset, policy = null) => [
-    ..._bitModStart(bits.opcodes.LSHIFT, 4),
-    ...bitOffset,
-    ...bitSize,
-    ...shift,
-    ..._uint((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit rshift operation.
- *
- * @param {Object} policy bit policy value.
- * @param {AerospikeExp} bitOffset Bit index of where to start operation.
- * @param {AerospikeExp} bitSize Number of bits to be operated on.
- * @param {number} shift Number of bits to shift by.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin resulting blob with the bytes operated on.
- */
-  rShift: (bin, shift, bitSize, bitOffset, policy = null) => [
-    ..._bitModStart(bits.opcodes.RSHIFT, 4),
-    ...bitOffset,
-    ...bitSize,
-    ...shift,
-    ..._uint((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit add operation.
- * Note: integers are stored big-endian.
- *
- * @param {Object} policy bit policy value.
- * @param {AerospikeExp} bitOffset Bit index of where to start operation.
- * @param {AerospikeExp} bitSize Number of bits to be operated on.
- * @param {AerospikeExp} value Integer expression for value to add.
- * @param {number} action bit overflow action value.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin resulting blob with the bytes operated on.
- */
-  add: (bin, action, value, bitSize, bitOffset, policy = null) => [
-    ..._bitModStart(bits.opcodes.ADD, 5),
-    ...bitOffset,
-    ...bitSize,
-    ...value,
-    ..._uint((policy ? policy.flags : 0)),
-    ..._uint(action),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit add operation.
- * Note: integers are stored big-endian.
- *
- * @param {Object} policy bit policy value.
- * @param {AerospikeExp} bitOffset Bit index of where to start operation.
- * @param {AerospikeExp} bitSize Number of bits to be operated on.
- * @param {AerospikeExp} value Integer expression for value to subtract.
- * @param {number} action bit overflow action value.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin resulting blob with the bytes operated on.
- */
-  subtract: (bin, action, value, bitSize, bitOffset, policy = null) => [
-    ..._bitModStart(bits.opcodes.SUBTRACT, 5),
-    ...bitOffset,
-    ...bitSize,
-    ...value,
-    ..._uint((policy ? policy.flags : 0)),
-    ..._uint(action),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit add operation.
- * Note: integers are stored big-endian.
- *
- * @param {Object} policy bit policy value.
- * @param {AerospikeExp} bitOffset Bit index of where to start operation.
- * @param {AerospikeExp} bitSize Number of bits to be operated on.
- * @param {AerospikeExp} value Integer expression for value to set.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin resulting blob with the bytes operated on.
- */
-  setInt: (bin, value, bitSize, bitOffset, policy = null) => [
-    ..._bitModStart(bits.opcodes.SET_INT, 4),
-    ...bitOffset,
-    ...bitSize,
-    ...value,
-    ..._uint((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /*********************************************************************************
- * BIT READ EXPRESSIONS
- *********************************************************************************/
-
-  /**
- * Create an expression that performs bit get operation.
- *
- * @param {AerospikeExp} bitOffset The bit index of where to start reading from.
- * @param {AerospikeExp} bitSize Number of bits to read from the blob bin.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} blob bin bit_size bits rounded up to the nearest byte size.
- */
-  get: (bin, bitSize, bitOffset) => [
-    ..._bitReadStart(exp.type.BLOB, bits.opcodes.GET, 2),
-    ...bitOffset,
-    ...bitSize,
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit count operation.
- *
- * @param {AerospikeExp} bitOffset The bit index of where to start reading from.
- * @param {AerospikeExp} bitSize Number of bits to read from the blob bin.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {number} integer value number of bits set to 1 in the bit_size region.
- */
-  count: (bin, bitSize, bitOffset) => [
-    ..._bitReadStart(exp.type.INT, bits.opcodes.COUNT, 2),
-    ...bitOffset,
-    ...bitSize,
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit lscan operation.
- *
- * @param {AerospikeExp} bitOffset The bit index of where to start reading from.
- * @param {AerospikeExp} bitSize Number of bits to read from the blob bin.
- * @param {AerospikeExp} value Boolean expression, true searches for 1, false for 0.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {number} integer value Index of the left most bit starting from __offset set to value.
- */
-  lScan: (bin, value, bitSize, bitOffset) => [
-    ..._bitReadStart(exp.type.INT, bits.opcodes.LSCAN, 3),
-    ...bitOffset,
-    ...bitSize,
-    ...value,
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit rscan operation.
- *
- * @param {AerospikeExp} bitOffset The bit index of where to start reading from.
- * @param {AerospikeExp} bitSize Number of bits to read from the blob bin.
- * @param {AerospikeExp} value Boolean expression, true searches for 1, false for 0.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {number} integer value Index of the right most bit starting from __offset set to value.
- */
-  rScan: (bin, value, bitSize, bitOffset) => [
-    ..._bitReadStart(exp.type.INT, bits.opcodes.RSCAN, 3),
-    ...bitOffset,
-    ...bitSize,
-    ...value,
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs bit get_int operation.
- *
- * @param {AerospikeExp} bitOffset The bit index of where to start reading from.
- * @param {AerospikeExp} bitSize Number of bits to read from the blob bin.
- * @param {boolean} sign Boolean value, true for signed, false for unsigned.
- * @param {AerospikeExp} bin A blob bin expression to apply this function to.
- * @return {AerospikeExp} integer value Index of the left most bit starting from offset set to value.
- */
-  getInt: (bin, sign, bitSize, bitOffset) => [
-    ..._bitReadStart(exp.type.INT, bits.opcodes.GET_INT, 3),
-    ...bitOffset,
-    ...bitSize,
-    ..._int((sign ? 1 : 0)),
-    ...bin
-  ]
-}
-
-/*********************************************************************************
- * HLL MODIFY EXPRESSIONS
- *********************************************************************************/
-
-const _hllModify = () => [
-  { op: exp.ops.CALL, count: 5 },
-  ..._rtype(exp.type.HLL),
-  ..._int(exp.sys.CALL_HLL | exp.sys.FLAG_MODIFY_LOCAL)
-]
-
-const _hllModifyStart = (op, param) => [
-  ..._hllModify(),
-  { op: exp.ops.CALL_VOP_START, count: 1 + param },
-  ..._int(op)
-]
-
-const _hllRead = (returnType) => [
-  { op: exp.ops.CALL, count: 5 },
-  ..._rtype(returnType),
-  ..._int(exp.sys.CALL_HLL)
-]
-
-const _hllReadStart = (returnType, op, param) => [
-  ..._hllRead(returnType),
-  { op: exp.ops.CALL_VOP_START, count: 1 + param },
-  ..._int(op)
-]
-
-exports.hll = {
-
-  /**
- * Create expression that creates a new HLL or resets an existing HLL with minhash bits.
- *
- * @param {Object} policy hll policy value.
- * @param {number} indexBitCount Number of index bits. Must be between 4 and 16 inclusive.
- * @param {number} mhBitCount Number of min hash bits. Must be between 4 and 51 inclusive.
- * @param {AerospikeExp} bin  A bin expression to apply this function to.
- * @return {AerospikeExp} hll bin Returns the resulting hll bin.
- */
-  initMH: (bin, mhBitCount, indexBitCount, policy = null) => [
-    ..._hllModifyStart(hlls.opcodes.INIT, 3),
-    ..._int(indexBitCount),
-    ..._int(mhBitCount),
-    ..._int((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create expression that creates a new HLL or resets an existing HLL.
- *
- * @param {Object} policy hll policy value.
- * @param {number} indexBitCount Number of index bits. Must be between 4 and 16 inclusive.
- * @param {AerospikeExp} bin  A bin expression to apply this function to.
- * @return {AerospikeExp} hll bin Returns the resulting hll bin.
- */
-  init: (bin, indexBitCount, policy = null) => [
-    ..._hllModifyStart(hlls.opcodes.INIT, 2),
-    ..._int(indexBitCount),
-    ..._int((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs operations hll addMh.
- *
- * @param {Object} policy hll policy value.
- * @param {AerospikeExp} list A list expression of elements to add to the HLL.
- * @param {number} indexBitCount Number of index bits. Must be between 4 and 16 inclusive.
- * @param {number} mhBitCount Number of min hash bits. Must be between 4 and 51 inclusive.
- * @param {AerospikeExp} bin  A bin expression to apply this function to.
- * @return {AerospikeExp} hll bin Returns the resulting hll bin after adding elements from list.
- */
-  addMH: (bin, mhBitCount, indexBitCount, list, policy = null) => [
-    ..._hllModifyStart(hlls.opcodes.ADD, 4),
-    ...list,
-    ..._int(indexBitCount),
-    ..._int(mhBitCount),
-    ..._int((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs operations hll add.
- *
- * @param {Object} policy hll policy value.
- * @param {AerospikeExp} list A list expression of elements to add to the HLL.
- * @param {number} indexBitCount Number of index bits. Must be between 4 and 16 inclusive.
- * @param {AerospikeExp} bin  A bin expression to apply this function to.
- * @return {AerospikeExp} hll bin Returns the resulting hll bin after adding elements from list.
- */
-  add: (bin, indexBitCount, list, policy = null) => [
-    ..._hllModifyStart(hlls.opcodes.ADD, 4),
-    ...list,
-    ..._int(indexBitCount),
-    ..._int(-1),
-    ..._int((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs operations hll update.
- *
- * @param {Object} policy hll policy value.
- * @param {AerospikeExp} list A list expression of elements to add to the HLL.
- * @param {AerospikeExp} bin A bin expression to apply this function to.
- * @return {AerospikeExp} hll bin Returns the resulting hll bin after adding elements from list.
- */
-  update: (bin, list, policy = null) => [
-    ..._hllModifyStart(hlls.opcodes.ADD, 4),
-    ...list,
-    ..._int(-1),
-    ..._int(-1),
-    ..._int((policy ? policy.flags : 0)),
-    ...bin
-  ],
-
-  /*********************************************************************************
- * HLL READ EXPRESSIONS
- *********************************************************************************/
-
-  /**
- * Create an expression that performs operations hll get count.
- *
- * @param {AerospikeExp} bin A bin expression to apply this function to.
- * @return {AerospikeExp} integer bin The estimated number of unique elements in an HLL.
- */
-  getCount: (bin) => [
-    ..._hllReadStart(exp.type.INT, hlls.opcodes.COUNT, 0),
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs operations hll get union.
- *
- * @param {AerospikeExp} list A list expression of HLLs to union with.
- * @param {AerospikeExp} bin A bin expression to apply this function to.
- * @return {AerospikeExp} hll bin HLL bin representing the set union.
- */
-  getUnion: (bin, list) => [
-    ..._hllReadStart(exp.type.HLL, hlls.opcodes.GET_UNION, 1),
-    ...list,
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs operations hll get union count.
- *
- * @param {AerospikeExp} list A list expression of HLLs to union with.
- * @param {AerospikeExp} bin A bin expression to apply this function to.
- * @return {AerospikeExp} integer bin Estimated number of elements in the set union.
- */
-  getUnionCount: (bin, list) => [
-    ..._hllReadStart(exp.type.INT, hlls.opcodes.UNION_COUNT, 1),
-    ...list,
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs operations hll get inersect count.
- *
- * @param {AerospikeExp} list A list expression of HLLs to intersect with.
- * @param {AerospikeExp} bin A bin expression to apply this function to.
- * @return {AerospikeExp} integer bin Estimated number of elements in the set intersection.
- */
-  getIntersectCount: (bin, list) => [
-    ..._hllReadStart(exp.type.INT, hlls.opcodes.INTERSECT_COUNT, 1),
-    ...list,
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs operations hll get similarity.
- *
- * @param {AerospikeExp} list A list expression of HLLs to calculate similarity with..
- * @param {AerospikeExp} bin A bin expression to apply this function to.
- * @return (float bin) Estimated similarity between 0.0 and 1.0.
- */
-  getSimilarity: (bin, list) => [
-    ..._hllReadStart(exp.type.FLOAT, hlls.opcodes.SIMILARITY, 1),
-    ...list,
-    ...bin
-  ],
-
-  /**
- * Create an expression that performs operations hll describe.
- *
- * @param {AerospikeExp} bin A bin expression to apply this function to.
- * @return {AerospikeExp} list bin A list containing the index_bit_count and minhash bit count.
- */
-  describe: (bin) => [
-    ..._hllReadStart(exp.type.LIST, hlls.opcodes.DESCRIBE, 0),
-    ...bin
-  ],
-
-  /**
- * Create an expression that checks if the HLL bin contains all keys in
- *  list..
- *
- * @param {AerospikeExp} list A list expression of keys to check if the HLL may contain them.
- * @param {AerospikeExp} bin A bin expression to apply this function to.
- * @return {AerospikeExp} integer bin 1 bin contains all of list, 0 otherwise.
- */
-  mayContain: (bin, list) => [
-    ..._hllReadStart(exp.type.INT, hlls.opcodes.MAY_CONTAIN, 1),
-    ...list,
-    ...bin
-  ]
-}
+ * The {@link module:aerospike/exp/lists|aerospike/exp/lists} module defines functions
+ * for expressions on the List datatype.
+ * 
+ * @summary List expressions.
+ */
+exports.lists = require('./exp_lists')
+
+/**
+ * The {@link module:aerospike/exp/maps|aerospike/exp/maps} module defines functions
+ * for expressions on the Map datatype.
+ * 
+ * @summary Map expressions.
+ */
+ exports.maps = require('./exp_maps')
+
+/**
+ * The {@link module:aerospike/exp/bit|aerospike/exp/bit} module defines functions
+ * for expressions on the Blob datatype
+ * 
+ * @summary Blob expressions.
+ */
+ exports.bit = require('./exp_bit')
+
+/**
+ * The {@link module:aerospike/exp/hll|aerospike/exp/hll} module defines functions
+ * for expressions on the HyperLogLog datatype
+ * 
+ * @summary HyperLogLog expressions.
+ */
+exports.hll = require('./exp_hll')
