+++ lib/exp_operations.js
@@ -17,39 +17,49 @@
 'use strict'
 
 /**
- * @module aerospike/exp_operations
+ * @module aerospike/exp/operations
  *
  * @description This module provides functions to easily create read/write operations
  * using Aerospike expressions to be performed on a record via
  * the {@link Client#operate} command.
  *
  * @see {@link Client#operate}
- * @see {@link Aerospike#exp}
+ * @see {@link module:aerospike/exp/operations|aerospike/exp/operations }
  *
  * @example
- *
  * const Aerospike = require('aerospike')
  * const op = Aerospike.operations
- * const exp = Aerospike.expressions
+ * const exp = Aerospike.exp
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
  * const tempBin = 'ExpVar' // this bin is to hold expression read operation output
  *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROSPIKE SERVER NODE HERE!
+ * const config = {  
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ *   
  * var ops = [
  *   op.append('a', 'xyz'),
  *   op.incr('b', 10),
  *   exp.operations.read(tempBin,
  *           exp.add(exp.binInt('b'), exp.binInt('b')),
  *          0),
+ *   op.read('a'),
  *   op.read('b')
  * ]
  *
- * Aerospike.connect((error, client) => {
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *   client.put(key, { a: 'abc', b: 42 }, (error) => {
  *     if (error) throw error
  *     client.operate(key, ops, (error, record) => {
  *       if (error) throw error
- *       console.log(record) // => { b: 52, ExpVar: 104 }
+ *       console.log(record.bins) // => { a: 'abcxyz', b: 52, ExpVar: 104 }
  *       client.close()
  *     })
  *   })
@@ -60,7 +70,8 @@ const as = require('bindings')('aerospike.node')
 const ops = as.expOperations
 
 /**
- * @class module:aerospike/exp_operations~ExpOperation
+ * @protected
+ * @class aerospike/exp_operations~ExpOperation
  *
  * @classdesc class for all expression operations executed with the {@link
  * Client#operate} command.
@@ -69,9 +80,7 @@ const ops = as.expOperations
  * * {@link module:aerospike/exp} - General expression on all types.
  */
 class ExpOperation {
-  /**
-   * @protected
-   */
+  
   constructor (op, bin, exp, flags, props) {
     this.op = op
     this.bin = bin
