+++ lib/filter.js
@@ -46,7 +46,7 @@ const GeoJSON = require('./geojson')
  */
 
 /**
- * @class module:aerospike/filter~SindexFilterPredicate
+ * @class aerospike/filter~SindexFilterPredicate
  * @classdesc SI filter predicate to limit the scope of a {@link Query}.
  *
  * Filter predicates must be instantiated using the methods in the {@link
