+++ lib/geojson.js
@@ -37,14 +37,27 @@
  * const Aerospike = require('aerospike')
  * const GeoJSON = Aerospike.GeoJSON
  * const Key = Aerospike.Key
- *
- * Aerospike.connect((error, client) => {
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *    
+ *   }
+ * }
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *   let key = new Key('test', 'demo', 'bob')
  *   let location = new GeoJSON({type: 'Point', coordinates: [103.913, 1.308]})
  *   client.put(key, {loc: location}, (error) => {
  *     if (error) throw error
- *     client.close()
+ *     client.get(key, (error, record) => {
+ *       if (error) throw error
+ *       console.log(record.bins.loc) // => {"type":"Point","coordinates":[103.913,1.308]}
+ *       client.close()
+ *     })
  *   })
  * })
  *
