+++ lib/hll.js
@@ -34,8 +34,16 @@
  * const Aerospike = require('aerospike')
  * const hll = Aerospike.hll
  * const key = new Aerospike.Key('test', 'demo', 'hllDemo')
- *
- * Aerospike.connect().then(async client => {
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * Aerospike.connect(config).then(async client => {
  *   const result = await client.operate(key, [
  *     hll.init('demo', 10),
  *     hll.add('demo', ['blue', 'green', 'red']),
@@ -43,7 +51,7 @@
  *     hll.add('demo', ['red', 'blue']),
  *     hll.getCount('demo')
  *   ])
- *   console.info('Count:', result.bins) // => Count: 5
+ *   console.log('Count:', result.bins.demo) // => Count: 5
  *   client.close()
  * })
  *
@@ -52,12 +60,19 @@
  * const Aerospike = require('aerospike')
  * const hll = Aerospike.hll
  * const ops = Aerospike.operations
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
  * const key1 = new Aerospike.Key('test', 'demo', 'hllDemo1')
  * const key2 = new Aerospike.Key('test', 'demo', 'hllDemo2')
  * const key3 = new Aerospike.Key('test', 'demo', 'hllDemo3')
  *
- * Aerospike.connect().then(async client => {
+ * Aerospike.connect(config).then(async client => {
  *   const { bins: { colors: colors1 } } = await client.operate(key1, [
  *     hll.add('colors', ['blue', 'green', 'orange', 'yellow'], 12),
  *     ops.read('colors')
@@ -75,7 +90,7 @@
  *     hll.getCount('colors')
  *   ])
  *
- *   console.info('Count:', count) // => Count: 9
+ *   console.log('Count:', count) // => Count: 9
  *   client.close()
  * })
  */
@@ -93,7 +108,7 @@ const GET_SIMILARITY_COMMAND = 54
 const DESCRIBE_COMMAND = 55
 
 /**
- * @class module:aerospike/hll~HLLOperation
+ * @class aerospike/hll~HLLOperation
  *
  * @classdesc Use the methods in the {@link module:aerospike/hll|hll}
  * module to create HLL operations for use with the {@link Client#operate}
