+++ lib/index_job.js
@@ -32,29 +32,51 @@ const util = require('util')
  * @example
  *
  * const Aerospike = require('aerospike')
- *
- * let binName = 'location'
- * let indexName = 'locationIndex'
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * 
+ * let binName = 'food'
+ * let indexName = 'foodIndex'
  * let options = {
- *   ns: 'test'
- *   set: 'demo'
+ *   ns: 'test',
+ *   set: 'demo',
  *   bin: binName,
  *   index: indexName,
- *   datatype: Aerospike.indexDataType.GEO2DSPHERE
+ *   datatype: Aerospike.indexDataType.STRING
  * }
  *
- * Aerospike.connect()
+ * Aerospike.connect(config)
  *   .then(client => {
- *     client.createIndex(options)
- *     .then(job => job.wait())
- *     .then(() => {
- *       console.info('secondary index (SI) %s on %s was created successfully', indexName, binName)
- *       client.close()
+ *     client.put(new Aerospike.Key('test', 'demo', 'mykey1'), {location: "Kale"})
+ *     .then((result) => {
+ *       client.createIndex(options)
+ *       .then(job => job.wait())
+ *       .then(() => {
+ *         console.info('secondary index (SI) %s on %s was created successfully', indexName, binName)
+ *         client.indexRemove('test', indexName)
+ *         .then(() => {
+ *           client.close()
+ *         })
+ *         .catch(error => {
+ *           console.error('Error removing index:', error)
+ *           client.close()
+ *         })
+ *       })
+ *       .catch(error => {
+ *         console.error('Error creating index:', error)
+ *         client.close()
+ *       })
  *     })
  *     .catch(error => {
- *       console.error('Error creating index:', error)
- *       client.close()
- *     })
+ *         console.error('Error writing record:', error)
+ *         client.close()
+ *       })
  *   })
  *   .catch(error => console.error('Error connecting to cluster:',  error))
  */
