+++ lib/info.js
@@ -32,7 +32,12 @@ const minimatch = require('minimatch')
  * @example
  *
  * const Aerospike = require('aerospike')
- * Aerospike.connect((error, client) => {
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ * }
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *
  *   var cmd = 'build\nfeatures'
@@ -232,7 +237,9 @@ const defaultSeparators = [smartParse]
  * Functions for dealing with specific info command results
  **********************************************************/
 
-/** Returns a new string with the last character removed */
+/** Returns a new string with the last character removed 
+ *  @private 
+ * */
 function chop (str) {
   return str.substring(0, str.length - 1)
 }
