+++ lib/lists.js
@@ -61,7 +61,7 @@ const AerospikeError = require('./error')
 const Operation = require('./operations').Operation
 
 /**
- * @class module:aerospike/lists~ListOperation
+ * @class aerospike/lists~ListOperation
  *
  * @classdesc Use the methods in the {@link module:aerospike/lists|lists}
  * module to create list operations for use with the {@link Client#operate}
@@ -82,15 +82,23 @@ class ListOperation extends Operation {
    * const Aerospike = require('aerospike')
    * const lists = Aerospike.lists
    * const key = new Aerospike.Key('test', 'demo', 'listsTest')
-   *
-   * Aerospike.connect().then(async client => {
+   * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+   * var config = {
+   *   hosts: '192.168.33.10:3000',
+   *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+   *   policies: {
+   *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+   *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+   *   }
+   * }
+   * Aerospike.connect(config).then(async client => {
    *   await client.put(key, { list: [32, 5, 85, 16, 22] })
    *   const ops = [
    *     lists.getByValueRange('list', 10, 30)
    *       .andReturn(lists.returnType.VALUE)
    *   ]
    *   const result = await client.operate(key, ops)
-   *   console.info('Result:', result.bins.list) // => Result: [ 16, 22 ]
+   *   console.log('Result:', result.bins.list) // => Result: [ 16, 22 ]
    *   client.close()
    * })
    */
@@ -114,15 +122,23 @@ class ListOperation extends Operation {
    * const Aerospike = require('aerospike')
    * const lists = Aerospike.lists
    * const key = new Aerospike.Key('test', 'demo', 'listsTest')
-   *
-   * Aerospike.connect().then(async (client) => {
+   * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+   * var config = {
+   *   hosts: '192.168.33.10:3000',
+   *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+   *   policies: {
+   *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+   *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+   *   }
+   * }
+   * Aerospike.connect(config).then(async (client) => {
    *   await client.put(key, { list: [[32, 5, 85], [16, 22]] })
    *   const ops = [
    *     lists.get('list', 0)
    *       .withContext((ctx) => ctx.addListIndex(1))
    *   ]
    *   const result = await client.operate(key, ops)
-   *   console.info('Result:', result.bins.list) // => Result: 16
+   *   console.log('Result:', result.bins.list) // => Result: 16
    *   client.close()
    * })
    *
@@ -132,8 +148,16 @@ class ListOperation extends Operation {
    * const lists = Aerospike.lists
    * const Context = Aerospike.cdt.Context
    * const key = new Aerospike.Key('test', 'demo', 'listsTest')
-   *
-   * Aerospike.connect().then(async (client) => {
+   * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+   * var config = {
+   *   hosts: '192.168.33.10:3000',
+   *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+   *   policies: {
+   *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+   *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+   *   }
+   * }
+   * Aerospike.connect(config).then(async (client) => {
    *   await client.put(key, { map: { nested: [32, 5, 85, 16, 22] } })
    *   const context = new Context().addMapKey('nested')
    *   const ops = [
@@ -141,7 +165,7 @@ class ListOperation extends Operation {
    *       .withContext(context)
    *   ]
    *   const result = await client.operate(key, ops)
-   *   console.info('Result:', result.bins.map) // => Result: 22
+   *   console.log('Result:', result.bins.map) // => Result: 22
    *   client.close()
    * })
    */
@@ -171,8 +195,16 @@ class ListOperation extends Operation {
    * const Aerospike = require('aerospike')
    * const lists = Aerospike.lists
    * const key = new Aerospike.Key('test', 'demo', 'listsTest')
-   *
-   * Aerospike.connect().then(async client => {
+   * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+   * var config = {
+   *   hosts: '192.168.33.10:3000',
+   *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+   *   policies: {
+   *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+   *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+   *   }
+   * }
+   * Aerospike.connect(config).then(async client => {
    *   await client.put(key, { tags: ['blue', 'yellow', 'pink'] })
    *   const ops = [
    *     lists.removeByValue('tags', 'yellow')
@@ -180,7 +212,7 @@ class ListOperation extends Operation {
    *   ]
    *   await client.operate(key, ops)
    *   const record = await client.get(key)
-   *   console.info('Result:', record.bins.tags) // => Result: [ 'yellow' ]
+   *   console.log('Result:', record.bins.tags) // => Result: [ 'yellow' ]
    *   client.close()
    * })
    */
@@ -207,7 +239,7 @@ class InvertibleListOp extends ListOperation {
 
 /**
  * @summary List order.
- * @description The order determines what kind of index the Aerospike server
+ * @description The order determines what kind of indices the Aerospike server
  * maintains for the list.
  *
  * @type Object
@@ -269,7 +301,7 @@ exports.writeFlags = as.lists.writeFlags
 exports.returnType = as.lists.returnType
 
 /**
- * @summary Sets the list order.
+ * @summary Sets the list order to <code>ORDERED</code> or <code>UNORDERED</code>
  * @description This operation does not return any result.
  *
  * @param {string} bin - The name of the bin. The bin must contain a List value.
@@ -313,22 +345,30 @@ exports.sort = function (bin, flags) {
  * @example
  *
  * const Aerospike = require('aerospike')
- * const op = Aerospike.operator
+ * const op = Aerospike.operations
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
  * var ops = [
  *   lists.append('tags', 'orange'),
  *   op.read('tags')
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
  *     client.operate(key, ops, (error, result) => {
  *       if (error) throw error
- *       console.log(result.tags) // => [ 'blue', 'yellow', 'pink', 'orange' ]
+ *       console.log(result.bins.tags) // => [ 'blue', 'yellow', 'pink', 'orange' ]
  *       client.close()
  *     })
  *   })
@@ -354,22 +394,30 @@ exports.append = function (bin, value, policy) {
  * @example
  *
  * const Aerospike = require('aerospike')
- * const op = Aerospike.operator
+ * const op = Aerospike.operations
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {                                                                                          
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
  * var ops = [
  *   lists.appendItems('tags', ['orange', 'green']),
  *   op.read('tags')
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
  *     client.operate(key, ops, (error, result) => {
  *       if (error) throw error
- *       console.log(result.tags) // => [ 'blue', 'yellow', 'pink', 'orange', 'green' ]
+ *       console.log(result.bins.tags) // => [ 'blue', 'yellow', 'pink', 'orange', 'green' ]
  *       client.close()
  *     })
  *   })
@@ -396,22 +444,31 @@ exports.appendItems = function (bin, list, policy) {
  * @example
  *
  * const Aerospike = require('aerospike')
- * const op = Aerospike.operator
+ * const op = Aerospike.operations
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * 
  * var ops = [
  *   lists.insert('tags', 2, 'orange'),
  *   op.read('tags')
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
  *     client.operate(key, ops, (error, result) => {
  *       if (error) throw error
- *       console.log(result.tags) // => [ 'blue', 'yellow', 'orange', 'pink' ]
+ *       console.log(result.bins.tags) // => [ 'blue', 'yellow', 'orange', 'pink' ]
  *       client.close()
  *     })
  *   })
@@ -438,22 +495,30 @@ exports.insert = function (bin, index, value, policy) {
  * @example
  *
  * const Aerospike = require('aerospike')
- * const op = Aerospike.operator
+ * const op = Aerospike.operations
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
  * var ops = [
  *   lists.insertItems('tags', 2, ['orange', 'green']),
  *   op.read('tags')
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
  *     client.operate(key, ops, (error, result) => {
  *       if (error) throw error
- *       console.log(result.tags) // => [ 'blue', 'yellow', 'orange', 'green', 'pink' ]
+ *       console.log(result.bins.tags) // => [ 'blue', 'yellow', 'orange', 'green', 'pink' ]
  *       client.close()
  *     })
  *   })
@@ -477,23 +542,33 @@ exports.insertItems = function (bin, index, list, policy) {
  * @example
  *
  * const Aerospike = require('aerospike')
- * const op = Aerospike.operator
+ * const op = Aerospike.operations
+ * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
  * var ops = [
  *   lists.pop('tags', 1)
  * ]
- *
- * Aerospike.client().connect((error, client) => {
+ * 
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
  *     client.operate(key, ops, (error, result) => {
  *       if (error) throw error
- *       console.log(result.tags) // => [ 'yellow' ]
+ *       console.log(result.bins.tags) // => yellow
  *       client.get(key, (error, record) => {
  *         if (error) throw error
- *         console.log(record) // => { tags: [ 'blue', 'pink' ] }
+ *         console.log(record.bins.tags) // => { [ 'blue', 'pink' ] }
  *         client.close()
  *       })
  *     })
@@ -507,7 +582,7 @@ exports.pop = function (bin, index) {
 }
 
 /**
- * @summary Removes and returns the list elements at the specified range.
+ * @summary Removes and returns the list elements in the specified range.
  *
  * @param {string} bin - The name of the bin. The bin must contain a List value.
  * @param {number} index - Index of the first element in the range.
@@ -519,21 +594,30 @@ exports.pop = function (bin, index) {
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
  * var ops = [
  *   lists.popRange('tags', 0, 2)
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
  *     client.operate(key, ops, (error, result) => {
  *       if (error) throw error
- *       console.log(result.tags) // => [ 'blue', 'yellow' ]
+ *       console.log(result.bins.tags) // => [ 'blue', 'yellow' ]
  *       client.get(key, (error, record) => {
  *         if (error) throw error
- *         console.log(record) // => { tags: [ 'pink' ] }
+ *         console.log(record.bins.tags) // => { [ 'pink' ] }
  *         client.close()
  *       })
  *     })
@@ -561,12 +645,21 @@ exports.popRange = function (bin, index, count) {
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
  * var ops = [
  *   lists.remove('tags', 1)
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
@@ -574,7 +667,7 @@ exports.popRange = function (bin, index, count) {
  *       if (error) throw error
  *       client.get(key, (error, record) => {
  *         if (error) throw error
- *         console.log(record) // => { tags: [ 'blue', 'pink' ] }
+ *         console.log(record.bins.tags) // => { [ 'blue', 'pink' ] }
  *         client.close()
  *       })
  *     })
@@ -588,7 +681,7 @@ exports.remove = function (bin, index) {
 }
 
 /**
- * @summary Removes the list elements at the specified range.
+ * @summary Removes the list elements in the specified range.
  * @description This operation returns the number of elements removed from the
  * list.
  *
@@ -602,12 +695,21 @@ exports.remove = function (bin, index) {
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
  * var ops = [
  *   lists.removeRange('tags', 0, 2)
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
@@ -615,7 +717,7 @@ exports.remove = function (bin, index) {
  *       if (error) throw error
  *       client.get(key, (error, record) => {
  *         if (error) throw error
- *         console.log(record) // => { tags: [ 'pink' ] }
+ *         console.log(record.bins.tags) // => { [ 'pink' ] }
  *         client.close()
  *       })
  *     })
@@ -653,17 +755,25 @@ exports.removeRange = function (bin, index, count) {
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'listsTest')
- *
- * Aerospike.connect().then(async client => {
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * Aerospike.connect(config).then(async client => {
  *   await client.put(key, { tags: ['blue', 'yellow', 'pink'] })
  *   const ops = [
  *     lists.removeByIndex('tags', 1)
  *       .andReturn(lists.returnType.VALUE)
  *   ]
  *   const result = await client.operate(key, ops)
- *   console.info('Result:', result.bins.tags) // => Result: yellow
+ *   console.log('Result:', result.bins.tags) // => Result: yellow
  *   const record = await client.get(key)
- *   console.info('Record:', record.bins.tags) // => Record: [ 'blue', 'pink' ]
+ *   console.log('Record:', record.bins.tags) // => Record: [ 'blue', 'pink' ]
  *   client.close()
  * })
  */
@@ -836,15 +946,24 @@ exports.removeByValueRange = function (bin, begin, end, returnType) {
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'listKey')
- *
- * Aerospike.connect().then(async client => {
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}), 
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * Aerospike.connect(config).then(async client => {
  *   await client.put(key, { list: [0, 4, 5, 9, 11, 15] })
  *   let result = await client.operate(key, [
  *     lists.removeByValueRelRankRange('list', 3, 3)
  *       .andReturn(lists.returnType.VALUE)])
- *   console.info(result.bins.list) // => [ 11, 15 ]
+ *   console.log(result.bins.list) // => [ 11, 15 ]
  *   let record = await client.get(key)
- *   console.info(record.bins.list) // => [ 0, 4, 5, 9 ]
+ *   console.log(record.bins.list) // => [ 0, 4, 5, 9 ]
  *   client.close()
  * })
  */
@@ -924,12 +1043,21 @@ exports.removeByRankRange = function (bin, rank, count, returnType) {
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
  * var ops = [
  *   lists.clear('tags')
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
@@ -937,7 +1065,7 @@ exports.removeByRankRange = function (bin, rank, count, returnType) {
  *       if (error) throw error
  *       client.get(key, (error, record) => {
  *         if (error) throw error
- *         console.log(record) // => { tags: [ ] }
+ *         console.log(record.bins.tags) // => { [ ] }
  *         client.close()
  *       })
  *     })
@@ -958,17 +1086,27 @@ exports.clear = function (bin) {
  * @param {ListPolicy} [policy] - Optional list policy.
  * @returns {Object} Operation that can be passed to the {@link Client#operate} command.
  *
- * @example
+ * @console.log
  *
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency varies with hardware selection. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout: 0}),
+ *     write : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout: 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout: 0})
+ *    
+ *    }
+ * }
  * var ops = [
  *   lists.set('tags', 1, 'green')
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
@@ -976,7 +1114,7 @@ exports.clear = function (bin) {
  *       if (error) throw error
  *       client.get(key, (error, record) => {
  *         if (error) throw error
- *         console.log(record) // => { tags: [ 'blue', 'green', 'pink' ] }
+ *         console.log(record.bins) // => { tags: [ 'blue', 'green', 'pink' ] }
  *         client.close()
  *       })
  *     })
@@ -988,7 +1126,7 @@ exports.set = function (bin, index, value, policy) {
 }
 
 /**
- * @summary Removes all list elements **not** within the specified range.
+ * @summary Removes all list elements that are not within the specified range.
  * @description This operation returns the number of list elements removed.
  *
  * @param {string} bin - The name of the bin. The bin must contain a List value.
@@ -1001,12 +1139,21 @@ exports.set = function (bin, index, value, policy) {
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0}) 
+ *   }
+ * }
  * var ops = [
  *   lists.trim('tags', 1, 1)
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
@@ -1014,7 +1161,7 @@ exports.set = function (bin, index, value, policy) {
  *       if (error) throw error
  *       client.get(key, (error, record) => {
  *         if (error) throw error
- *         console.log(record) // => { tags: [ 'yellow' ] }
+ *         console.log(record.bins.tags) // => { ['yellow'] }
  *         client.close()
  *       })
  *     })
@@ -1040,22 +1187,28 @@ exports.trim = function (bin, index, count) {
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0}) 
+ *   }
+ * }
  * var ops = [
  *   lists.get('tags', 0)
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
- *     client.operate(key, ops, (error) => {
+ *     client.operate(key, ops, (error, result) => {
  *       if (error) throw error
- *       client.get(key, (error, record) => {
- *         if (error) throw error
- *         console.log(record) // => { tags: 'blue' }
- *         client.close()
- *       })
+ *       console.log(result.bins) // => { tags: 'blue' }
+ *       client.close()
+ *       
  *     })
  *   })
  * })
@@ -1067,7 +1220,7 @@ exports.get = function (bin, index) {
 }
 
 /**
- * @summary Returns the list element at the specified range.
+ * @summary Returns the list element in the specified range.
  *
  * @param {string} bin - The name of the bin. The bin must contain a List value.
  * @param {number} index - Index of the first element in the range.
@@ -1079,22 +1232,28 @@ exports.get = function (bin, index) {
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
  * var ops = [
  *   lists.getRange('tags', 1)
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
- *     client.operate(key, ops, (error) => {
+ *     client.operate(key, ops, (error, result) => {
  *       if (error) throw error
- *       client.get(key, (error, record) => {
- *         if (error) throw error
- *         console.log(record) // => { tags: [ 'yellow', 'pink' ] }
- *         client.close()
- *       })
+ *       console.log(result.bins.tags) // => { [ 'yellow', 'pink' ] }
+ *       client.close()
+ *       
  *     })
  *   })
  * })
@@ -1107,7 +1266,7 @@ exports.getRange = function (bin, index, count) {
 }
 
 /**
- * @summary Retrieves a single list element identified by its index from the list.
+ * @summary Retrieves a single list element from the list using a specified index.
  * @description This operation returns the data specified by <code>returnType</code>.
  *
  * @param {string} bin - The name of the bin. The bin must contain a List value.
@@ -1130,15 +1289,23 @@ exports.getRange = function (bin, index, count) {
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'listsTest')
- *
- * Aerospike.connect().then(async client => {
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * Aerospike.connect(config).then(async client => {
  *   await client.put(key, { tags: ['blue', 'yellow', 'pink'] })
  *   const ops = [
  *     lists.getByIndex('tags', 1)
  *       .andReturn(lists.returnType.VALUE)
  *   ]
  *   const result = await client.operate(key, ops)
- *   console.info('Result:', result.bins.tags) // => Result: yellow
+ *   console.log('Result:', result.bins.tags) // => Result: yellow
  *   client.close()
  * })
  */
@@ -1311,14 +1478,22 @@ exports.getByValueRange = function (bin, begin, end, returnType) {
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'listKey')
- *
- * Aerospike.connect().then(async client => {
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * Aerospike.connect(config).then(async client => {
  *   await client.put(key, { list: [0, 4, 5, 9, 11, 15] })
  *   await client.operate(key, [ lists.setOrder('list', lists.order.ORDERED) ])
  *   let result = await client.operate(key, [
  *     lists.getByValueRelRankRange('list', 5, -1, 2)
  *       .andReturn(lists.returnType.VALUE)])
- *   console.info(result.bins.list) // => [ 4, 5 ]
+ *   console.log(result.bins.list) // => [ 4, 5 ]
  *   client.close()
  * })
  */
@@ -1407,17 +1582,28 @@ exports.getByRankRange = function (bin, rank, count, returnType) {
  * var ops = [
  *   lists.increment('counters', 1, 3)
  * ]
- *
- * Aerospike.client().connect((error, client) => {
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * 
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { counters: [1, 2, 3] }, (error) => {
  *     if (error) throw error
  *     client.operate(key, ops, (error, result) => {
  *       if (error) throw error
- *       console.log(result['counters']) => 5
+ *       console.log(result['bins']['counters']) // => 5
  *       client.get(key, (error, record) => {
  *         if (error) throw error
- *         console.log(record) // => { counters: [1, 5, 3] }
+ *         console.log(record['bins']['counters']) // => { [1, 5, 3] }
  *         client.close()
  *       })
  *     })
@@ -1443,18 +1629,26 @@ exports.increment = function (bin, index, value, policy) {
  * const Aerospike = require('aerospike')
  * const lists = Aerospike.lists
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
  * var ops = [
  *   lists.size('tags')
  * ]
  *
- * Aerospike.client().connect((error, client) => {
+ * Aerospike.client(config).connect((error, client) => {
  *   if (error) throw error
  *   client.put(key, { tags: ['blue', 'yellow', 'pink'] }, (error) => {
  *     if (error) throw error
- *     client.operate(key, ops, (error) => {
+ *     client.operate(key, ops, (error, result) => {
  *       if (error) throw error
- *       console.log(record) // => { tags: 3 }
+ *       console.log(result.bins.tags) // => { 3 }
  *       client.close()
  *     })
  *   })
