+++ lib/maps.js
@@ -81,7 +81,18 @@ const Operation = require('./operations').Operation
  * const maps = Aerospike.maps
  * const key = new Aerospike.Key('test', 'demo', 'mapKey')
  *
- * Aerospike.connect().then(async client => {
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     remove : new Aerospike.RemovePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * 
+ * Aerospike.connect(config).then(async client => {
  *   let ops = [
  *     maps.put('map', 'e', 5, {
  *       order: maps.order.KEY_ORDERED
@@ -106,7 +117,7 @@ const Operation = require('./operations').Operation
  */
 
 /**
- * @class module:aerospike/maps~MapOperation
+ * @class aerospike/maps~MapOperation
  *
  * @classdesc Use the methods in the {@link module:aerospike/maps|maps}
  * module to create map operations for use with the {@link Client#operate}
@@ -128,8 +139,17 @@ class MapOperation extends Operation {
    * const Aerospike = require('aerospike')
    * const maps = Aerospike.maps
    * const key = new Aerospike.Key('test', 'demo', 'mapKey')
-   *
-   * Aerospike.connect().then(async client => {
+   * 
+   * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+   * var config = {
+   *   hosts: '192.168.33.10:3000',
+   *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+   *   policies: {
+   *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+   *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+   *   }
+   * }
+   * Aerospike.connect(config).then(async client => {
    *   await client.put(key, { map: { a: 1, b: 2, c: 3, d: 4, e: 5 } })
    *   const ops = [
    *     maps.removeByKeyRange('map', 'b', 'd')
@@ -160,8 +180,16 @@ class MapOperation extends Operation {
    * const Aerospike = require('aerospike')
    * const maps = Aerospike.maps
    * const key = new Aerospike.Key('test', 'demo', 'mapsTest')
-   *
-   * Aerospike.connect().then(async (client) => {
+   * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+   * var config = {
+   *   hosts: '192.168.33.10:3000',
+   *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+   *   policies: {
+   *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+   *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0}) 
+   *   }
+   * }
+   * Aerospike.connect(config).then(async (client) => {
    *   await client.put(key, { maps: [{ a: 1, b: 2 }, { b: 3, c: 4 }] })
    *   const ops = [
    *     maps.getByKey('maps', 'b')
@@ -179,8 +207,16 @@ class MapOperation extends Operation {
    * const maps = Aerospike.maps
    * const Context = Aerospike.cdt.Context
    * const key = new Aerospike.Key('test', 'demo', 'listsTest')
-   *
-   * Aerospike.connect().then(async (client) => {
+   * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+   * var config = {
+   *   hosts: '192.168.33.10:3000',
+   *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+   *   policies: {
+   *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+   *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+   *   }
+   * }
+   * Aerospike.connect(config).then(async (client) => {
    *   await client.put(key, { maps: { alpha: { a: 1, b: 2 }, beta: { b: 3, c: 4 } } })
    *   const context = new Context().addMapKey('alpha')
    *   const ops = [
@@ -220,10 +256,7 @@ exports.order = as.maps.order
 /**
  * @summary Map write mode.
  *
- * @description The write mode determines whether a write operation succeeds,
- * depending on whether the map key(s) to be written already exist. It also
- * determines whether a new map will be created automatically if the record
- * bin, which the map operation is targeting, is currently empty.
+ * @description Write mode is used to determine the criteria for a successful operation.
  *
  * Map write mode should only be used for server versions prior to v4.3. For
  * server versions v4.3 or later, the use of {@link
@@ -245,10 +278,7 @@ exports.writeMode = as.maps.writeMode
 /**
  * @summary Map write flags.
  *
- * @description The write flags determine whether a write operation succeeds,
- * depending on whether the map key(s) to be written already exist. They also
- * determine whether a new map will be created automatically if the record bin,
- * which the map operation is targeting, is currently empty.
+ * @description Write flags are used to determine the criteria for a successful operation.
  *
  * Map write flags require server version v4.3 or later. For earier server
  * versions, set the {@link module:aerospike/maps.writeMode|writeMode} instead.
@@ -341,7 +371,7 @@ exports.put = function (bin, key, value, policy) {
 /**
  * @summary Writes each entry of the given map to the map bin on the server.
  *
- * @description Depending on the map policy and whether an entry with the same
+ * @description For each item, depending on the map policy and whether an entry with the same
  * key already exists in the map, a new entry will be added to the map or the
  * existing entry with the same key will be updated. If the bin does not yet
  * contain a map value, a new map may be created.
@@ -556,8 +586,17 @@ exports.removeByKeyRange = function (bin, begin, end, returnType) {
  * const Aerospike = require('aerospike')
  * const maps = Aerospike.maps
  * const key = new Aerospike.Key('test', 'demo', 'mapKey')
- *
- * Aerospike.connect()
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * Aerospike.connect(config)
  *   .then(async client => {
  *     await client.put(key, { map: { a: 17, e: 2, f: 15, j: 10 } })
  *     let result = await client.operate(key, [
@@ -695,7 +734,17 @@ exports.removeByValueRange = function (bin, begin, end, returnType) {
  * const maps = Aerospike.maps
  * const key = new Aerospike.Key('test', 'demo', 'mapKey')
  *
- * Aerospike.connect()
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *   }
+ * }
+ * Aerospike.connect(config)
  *   .then(async client => {
  *     await client.put(key, { map: { e: 2, j: 10, f: 15, a: 17 } })
  *     let result = await client.operate(key, [
@@ -925,8 +974,16 @@ exports.getByKeyRange = function (bin, begin, end, returnType) {
  * const Aerospike = require('aerospike')
  * const maps = Aerospike.maps
  * const key = new Aerospike.Key('test', 'demo', 'mapKey')
- *
- * Aerospike.connect()
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0}) 
+ *   }
+ * }
+ * Aerospike.connect(config)
  *   .then(async client => {
  *     await client.put(key, { map: { a: 17, e: 2, f: 15, j: 10 } })
  *     let result = await client.operate(key, [
@@ -1037,8 +1094,17 @@ exports.getByValueRange = function (bin, begin, end, returnType) {
  * const Aerospike = require('aerospike')
  * const maps = Aerospike.maps
  * const key = new Aerospike.Key('test', 'demo', 'mapKey')
- *
- * Aerospike.connect()
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0})
+ *    
+ *   }
+ * }
+ * Aerospike.connect(config)
  *   .then(async client => {
  *     await client.put(key, { map: { e: 2, j: 10, f: 15, a: 17 } })
  *     let result = await client.operate(key, [
