+++ lib/operations.js
@@ -29,45 +29,50 @@
  * const Aerospike = require('aerospike')
  * const op = Aerospike.operations
  * const key = new Aerospike.Key('test', 'demo', 'mykey1')
- *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     operate : new Aerospike.OperatePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}), 
+ *   }
+ * }
  * var ops = [
  *   op.append('a', 'xyz'),
  *   op.incr('b', 10),
  *   op.read('b')
  * ]
  *
- * Aerospike.connect((error, client) => {
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *   client.put(key, { a: 'abc', b: 42 }, (error) => {
  *     if (error) throw error
  *     client.operate(key, ops, (error, record) => {
  *       if (error) throw error
- *       console.log(record) // => { b: 52 }
+ *       console.log(record.bins) // => { b: 52 }
  *       client.close()
  *     })
  *   })
  * })
  */
-
 const as = require('bindings')('aerospike.node')
 const ops = as.scalarOperations
-
 /**
- * @class module:aerospike/operations~Operation
- *
- * @classdesc Base class for all operations executed with the {@link
- * Client#operate} command.
- *
- * Operations can be created using the methods in one of the following modules:
- * * {@link module:aerospike/operations} - General operations on all types.
- * * {@link module:aerospike/lists} - Operations on CDT List values.
- * * {@link module:aerospike/maps} - Operations on CDT Map values.
- * * {@link module:aerospike/bitwise} - Operations on Bytes values.
- */
+  * @class aerospike/operations~Operation
+  * @protected
+  * @classdesc Base class for all operations executed with the {@link
+  * Client#operate} command.
+  *
+  * Operations can be created using the methods in one of the following modules:
+  * * {@link module:aerospike/operations} - General operations on all types.
+  * * {@link module:aerospike/lists} - Operations on CDT List values.
+  * * {@link module:aerospike/maps} - Operations on CDT Map values.
+  * * {@link module:aerospike/bitwise} - Operations on Bytes values.
+  * 
+  */
 class Operation {
-  /**
-   * @protected
-   */
+  
   constructor (op, bin, props) {
     this.op = op
     this.bin = bin
@@ -108,7 +113,7 @@ exports.write = function (bin, value) {
  * value must be of the same type.
  *
  * @param {string} bin - The name of the bin.
- * @param {(number|Object)} value - The number|{@link Double} value to increment the bin by.
+ * @param {(number|Object)} value - The <code>number</code>|{@link Double} value to increment the bin by.
  * @returns {Operation} Operation that can be passed to the {@link Client#operate} command.
  */
 exports.add = function (bin, value) {
