+++ lib/policy.js
@@ -100,16 +100,8 @@ const MapPolicy = require('./policies/map_policy')
  *
  * const Aerospike = require('aerospike')
  *
- * // Set default policies for read and write commands
- * const defaults = {
- *   socketTimeout: 50,
- *   totalTimeout: 100
- * }
  * const config = {
- *   policies: {
- *     read: new Aerospike.ReadPolicy(defaults),
- *     write: new Aerospike.WritePolicy(defaults)
- *   }
+ *   hosts: '192.168.33.10:3000'   
  * }
  *
  * const key = new Aerospike.Key('test', 'demo', 'k1')
@@ -121,7 +113,9 @@ const MapPolicy = require('./policies/map_policy')
  *     // Override policy for put command
  *     let policy = new Aerospike.policy.WritePolicy({
  *       exists: Aerospike.policy.exists.CREATE,
- *       key: Aerospike.policy.key.SEND
+ *       key: Aerospike.policy.key.SEND,
+ *       socketTimeout: 0,
+ *       totalTimeout: 0
  *     })
  *
  *     return client.put(key, record, {}, policy)
@@ -158,8 +152,15 @@ const MapPolicy = require('./policies/map_policy')
  *
  * const Aerospike = require('aerospike')
  * const key = new Aerospike.Key('test', 'test', 'myKey')
- *
- * Aerospike.connect().then(async (client) => {
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}) 
+ *   }
+ * }
+ * Aerospike.connect(config).then(async (client) => {
  *   await client.put(key, { foo: 'bar' })
  *
  *   const record = await client.get(key)
@@ -289,7 +290,7 @@ exports.readModeSC = as.policy.readModeSC
 exports.commitLevel = as.policy.commitLevel
 
 /**
- * A policy affecting the behavior of most operations.
+ * A base class extended to client policies. 
  *
  * @summary {@link BasePolicy} class
  */
