+++ lib/record.js
@@ -20,16 +20,27 @@
  * @class Record
  * @classdesc Aerospike Record
  *
- * @summary Construct a new Aerospike Record instance.
+ *
  *
  * A record with the Aerospike database consists of one or more record "bins"
  * (name-value pairs) and meta-data, including time-to-live and generation; a
  * record is uniquely identified by it's key within a given namespace.
+ * 
+ * @summary Construct a new Aerospike Record instance.
  *
  * @example <caption>Writing a new record with 5 bins while setting a record TTL.</caption>
  *
  * const Aerospike = require('aerospike')
- *
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *    }
+ * }
+ * 
  * let bins = {
  *   int: 123,
  *   double: 3.1415,
@@ -43,14 +54,25 @@
  * }
  * let key = new Aerospike.Key('test', 'demo', 'myKey')
  *
- * Aerospike.connect()
+ * Aerospike.connect(config)
  *   .then(client => {
  *     return client.put(key, bins, meta)
- *       .then(() => client.close())
+ *     .then(() => {
+ *       client.get(key)
+ *       .then((record) => {
+ *         console.log(record)
+ *         client.close()
+ *       }) 
  *       .catch(error => {
+ *         console.log(record)
  *         client.close()
  *         return Promise.reject(error)
  *       })
+ *     })
+ *     .catch(error => {
+ *       client.close()
+ *       return Promise.reject(error)
+ *     })
  *   })
  *   .catch(error => console.error('Error:', error))
  *
@@ -58,11 +80,23 @@
  *
  * const Aerospike = require('aerospike')
  *
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *    }
+ * }
+ * 
  * let key = new Aerospike.Key('test', 'demo', 'myKey')
  *
- * Aerospike.connect()
+ * Aerospike.connect(config)
  *   .then(client => {
- *     client.get(key)
+ *     client.put(key, {tags : ['blue', 'pink']})
+ *     .then(() => {
+ *       client.get(key)
  *       .then(record => {
  *         console.info('Key:', record.key)
  *         console.info('Bins:', record.bins)
@@ -73,8 +107,13 @@
  *       .catch(error => {
  *         client.close()
  *         return Promise.reject(error)
- *       })
- *   })
+ *       }) 
+ *     })
+ *     .catch(error => {
+ *       client.close()
+ *       return Promise.reject(error)
+ *     })
+ *   }) 
  *   .catch(error => console.error('Error:', error))
  *
  * @since v5.0.0
@@ -82,25 +121,41 @@
  * @example <caption>Fetching a batch of records.</caption>
  *
  * const Aerospike = require('aerospike')
+ * const op = Aerospike.operations
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     batch : new Aerospike.BatchPolicy({socketTimeout : 0, totalTimeout : 0})
+ *    
+ *    }
+ * }
+ * 
  * var batchRecords = [
- *   { type: batchType.BATCH_READ,
+ *   { type: Aerospike.batchType.BATCH_READ,
  *     key: new Aerospike.Key('test', 'demo', 'key1'), bins: ['i', 's'] },
- *   { type: batchType.BATCH_READ,
+ *   { type: Aerospike.batchType.BATCH_READ,
  *     key: new Aerospike.Key('test', 'demo', 'key2'), readAllBins: true },
- *   { type: batchType.BATCH_READ,
+ *   { type: Aerospike.batchType.BATCH_READ,
  *     key: new Aerospike.Key('test', 'demo', 'key3'),
  *     ops:[
  *          op.read('blob-bin')
  *         ]}
  * ]
- * Aerospike.connect(function (error, client) {
+ * Aerospike.connect(config, function (error, client) {
  *   if (error) throw error
  *   client.batchRead(batchRecords, function (error, results) {
  *     if (error) throw error
  *     results.forEach(function (result) {
  *       console.log(result)
+ *       
  *     })
+ *     client.close()
  *   })
+ *   
  * })
  *
  * @since v5.0.0
@@ -108,6 +163,18 @@
  * @example <caption>Applying  functions on batch of records.</caption>
  *
  * const Aerospike = require('aerospike')
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     read : new Aerospike.ReadPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *     write : new Aerospike.WritePolicy({socketTimeout : 0, totalTimeout : 0}),
+ *    
+ *    }
+ * }
+ * 
  * const batchType = Aerospike.batchType
  * var batchRecords = [
  *   { type: batchType.BATCH_READ,
@@ -145,7 +212,7 @@
  *          }
  *    }
  * ]
- * Aerospike.connect(function (error, client) {
+ * Aerospike.connect(config, function (error, client) {
  *   if (error) throw error
  *   client.batchApply(batchRecords, udf, function (error, results) {
  *     if (error) throw error
