+++ lib/record_stream.js
@@ -40,7 +40,17 @@ const inherits = require('util').inherits
  * @example
  *
  * const Aerospike = require('aerospike')
- * Aerospike.connect((error, client) => {
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     scan : new Aerospike.ScanPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *    }
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *   var recordsSeen = 0
  *   var scan = client.scan('test', 'demo')
