+++ lib/scan.js
@@ -80,7 +80,16 @@ const RecordStream = require('./record_stream')
  *
  * const Aerospike = require('aerospike')
  *
- * Aerospike.connect((error, client) => {
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     scan : new Aerospike.ScanPolicy({socketTimeout : 0, totalTimeout : 0}),
+ *    }
+ * }
+ * 
+ * Aerospike.connect(config, (error, client) => {
  *   if (error) throw error
  *
  *   const scan = client.scan('test', 'demo')
@@ -259,8 +268,17 @@ Scan.prototype.background = function (udfModule, udfFunction, udfArgs, policy, s
  * @example <caption>Increment count bin on all records in set using a background scan</caption>
  *
  * const Aerospike = require('aerospike')
- *
- * Aerospike.connect().then(async (client) => {
+ * 
+ * // INSERT HOSTNAME AND PORT NUMBER OF AEROPSIKE SERVER NODE HERE!
+ * var config = {
+ *   hosts: '192.168.33.10:3000',
+ *   // Timeouts disabled, latency dependent on server location. Configure as needed.
+ *   policies: {
+ *     scan : new Aerospike.ScanPolicy({socketTimeout : 0, totalTimeout : 0})
+ *    }
+ * }
+ * 
+ * Aerospike.connect(config).then(async (client) => {
  *   const scan = client.scan('namespace', 'set')
  *   const ops = [Aerospike.operations.incr('count', 1)]
  *   const job = await scan.operate(ops)
