+++ lib/typedefs.js
@@ -224,23 +224,6 @@
  * @since v4.0.0
  */
 
-/**
- * @typedef {object} ApplyPolicy
- * @typedef {object} BatchPolicy
- * @typedef {object} InfoPolicy
- * @typedef {object} OperatePolicy
- * @typedef {object} ReadPolicy
- * @typedef {object} RemovePolicy
- * @typedef {object} ScanPolicy
- * @typedef {object} QueryPolicy
- * @typedef {object} WritePolicy
- * @typedef {object} BitwisePolicy
- * @typedef {object} MapPolicy
- * @typedef {object} CommandQueuePolicy
- *
- * @summary Aerospike Policy objects
- *
- */
 
 /**
  * @typedef {Object} Policies
@@ -254,6 +237,9 @@
  * @property {ScanPolicy} scan - Default scan policy
  * @property {QueryPolicy} query - Default query policy
  * @property {WritePolicy} write - Default write policy
+ * @property {MapPolicy} map - Default map policy
+ * @property {ListPolicy} list - Default list policy
+ * @property {CommandQueuePolicy} commandQueue - Default command queue policy
  */
 
 /**
@@ -263,19 +249,7 @@
  *
  */
 
-/**
- * @typedef {object} Client
- *
- * @summary Aerospike Client object
- *
- */
 
-/**
- * @typedef {object} Key
- *
- * @summary Aerospike Key object
- *
- */
 
 /**
  * @typedef {object} RecordObject
